% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MarxanResults.R, R/MarxanSolved.R
\name{amountheld}
\alias{amountheld}
\alias{amountheld.MarxanResults}
\alias{amountheld.MarxanSolved}
\title{Extract amount held for a solution}
\usage{
amountheld(x, ...)

\method{amountheld}{MarxanResults}(x, y = NULL)

\method{amountheld}{MarxanSolved}(x, y = NULL)
}
\arguments{
\item{x}{"MarxanResults" or "MarxanSolved" object.}

\item{...}{not used.}

\item{y}{"NULL" to return all values, "integer" 0 to return values for best solution, "integer" value greater than 0 for \code{y}'th solution value.}
}
\value{
"matrix" or "numeric" vector depending on arguments.
}
\description{
This function returns the amount held for each species in a solution.
}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved}}, \code{\link{marxan}}
}

