% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MarxanData.R, R/MarxanResults.R, R/generics.R
\docType{methods}
\name{cache,MarxanData,character,ANY-method}
\alias{cache}
\alias{cache,MarxanData,character,ANY-method}
\alias{cache,MarxanData,character,missing-method}
\alias{cache,MarxanResults,character,ANY-method}
\alias{cache,MarxanResults,character,missing-method}
\title{Get and Set Cache Methods}
\usage{
\S4method{cache}{MarxanData,character,ANY}(x, name, y)

\S4method{cache}{MarxanData,character,missing}(x, name, y)

\S4method{cache}{MarxanResults,character,ANY}(x, name, y)

\S4method{cache}{MarxanResults,character,missing}(x, name, y)

cache(x, name, y)
}
\arguments{
\item{x}{"MarxanData" or "MarxanResults" object}

\item{name}{"character" hash.}

\item{y}{if "ANY" this object gets cached with name, else if "missing" the object hashed at name gets returned.}
}
\value{
"any" or "NULL" depends on y argument.
}
\description{
Getter and setter methods for caches in MarxanData and MarxanResults object.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = MarxanData,name = character,y = ANY}: 

\item \code{x = MarxanData,name = character,y = missing}: 

\item \code{x = MarxanResults,name = character,y = ANY}: 

\item \code{x = MarxanResults,name = character,y = missing}: 
}}
\note{
caches are implemented using environments, the hash is used as the name of the object in the environment.
}
\keyword{internal}

