% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcBoundaryData.R
\docType{methods}
\name{calcBoundaryData,PolySet-method}
\alias{calcBoundaryData,PolySet-method}
\title{describeIn calcBoundaryData}
\usage{
\S4method{calcBoundaryData}{PolySet}(x, tolerance = 0.001, lengthFactor = 1,
  edgeFactor = 1)
}
\description{
describeIn calcBoundaryData
}

