% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcBoundaryData.R
\name{calcBoundaryData.SpatialPolygons}
\alias{calcBoundaryData.SpatialPolygons}
\title{describeIn calcBoundaryData}
\usage{
\method{calcBoundaryData}{SpatialPolygons}(x, tolerance = 0.001,
  lengthFactor = 1, edgeFactor = 1, ...)
}
\description{
describeIn calcBoundaryData
}

