% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPuVsSpeciesData.R
\name{calcPuVsSpeciesData}
\alias{calcPuVsSpeciesData}
\alias{calcPuVsSpeciesData.SpatialPolygons}
\alias{calcPuVsSpeciesData.SpatialPolygonsDataFrame}
\title{Calculate planning units vs. species data}
\usage{
calcPuVsSpeciesData(x, ...)

\method{calcPuVsSpeciesData}{SpatialPolygons}(x, y, ids = seq_len(nlayers(y)),
  ncores = 1, gdal = FALSE, ...)

\method{calcPuVsSpeciesData}{SpatialPolygonsDataFrame}(x, y,
  ids = seq_len(nlayers(y)), ncores = 1, gdal = FALSE, field = NULL,
  ...)
}
\arguments{
\item{x}{"SpatialPolygons" or "SpatialPolygonsDataFrame" object.}

\item{...}{not used.}

\item{y}{"RasterLayer", "RasterStack", or "RasterBrick" object.}

\item{ids}{"integer" vector of ids. Defaults to indices of layers in the \code{y}.}

\item{ncores}{"integer" Number of cores to use for processing. Defaults to 1.}

\item{gdal}{"logical" Should raster processing be performed using GDAL libraries? Defaults to \code{FALSE}.}

\item{field}{"character" "integer" index or "character" name of column with planning unit ids. Valid only for "SpatialPolygonsDataFrame" objects. Default behaviour is to treat each polygon as a different planning unit.}
}
\value{
data.frame with sum of raster values in each polygon.
}
\description{
This function calculates the sum of species values in each planning unit.
Be aware that using polygons with overlaps will result in inaccuracies.
By default all polygons will be treated as having separate ids.
}
\examples{
data(taspu, tasinvis)
puvspr1.dat<-calcPuVsSpeciesData(taspu, tasinvis[[1]])
puvspr2.dat<-calcPuVsSpeciesData(taspu, tasinvis)
}
\seealso{
\code{\link{is.gdalInstalled}}, \code{\link{zonalSum}}, \url{http://www.gdal.org/}, \url{http://trac.osgeo.org/gdal/wiki/DownloadingGdalBinaries}.
}

