% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R, R/MarxanSolved.R, R/MarxanUnsolved.R, R/generics.R
\name{costs.MarxanData}
\alias{costs}
\alias{costs.MarxanData}
\alias{costs.MarxanSolved}
\alias{costs.MarxanUnsolved}
\alias{costs<-.MarxanData}
\alias{costs<-.MarxanSolved}
\alias{costs<-.MarxanUnsolved}
\title{Planning unit costs}
\usage{
\method{costs}{MarxanData}(x)

\method{costs}{MarxanData}(x) <- value

\method{costs}{MarxanSolved}(x)

\method{costs}{MarxanSolved}(x) <- value

\method{costs}{MarxanUnsolved}(x)

\method{costs}{MarxanUnsolved}(x) <- value

costs(x, ...)
}
\arguments{
\item{x}{any "MarxanOpts", "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.}
}
\description{
This function returns or assigns the planing unit costs for a Marxan object.
}
\section{Methods (by class)}{
\itemize{
\item \code{MarxanData}: 

\item \code{MarxanData}: 

\item \code{MarxanSolved}: 

\item \code{MarxanSolved}: 

\item \code{MarxanUnsolved}: 

\item \code{MarxanUnsolved}: 
}}
\note{
This generic method does not work on "MaranResults" objects because they do not store this information.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

