% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MarxanResults.R, R/MarxanSolved.R
\name{dendrogram}
\alias{dendrogram}
\alias{dendrogram.MarxanResults}
\alias{dendrogram.MarxanSolved}
\title{Dendrogram of Marxan solutions}
\usage{
dendrogram(x, ...)

\method{dendrogram}{MarxanResults}(x, type = "mds", var = "selections",
  nbest = 1, ..., force_reset = FALSE)

\method{dendrogram}{MarxanSolved}(x, type = "mds", var = "selections",
  nbest = 1, ..., force_reset = FALSE)
}
\arguments{
\item{x}{"MarxanResults" or "MarxanSolved" object.}

\item{...}{arguments to \code{\link[vegan]{monoMDS}}, \code{\link[stats]{prcomp}}, \code{\link{mds.MarxanResults}}, and \code{\link{pca.MarxanResults}}.}

\item{type}{"character" use 'nmds' or 'pca' or 'dist' for analysis?}

\item{var}{"character" should solutions be compared based on selections ('selections'), or the amount held ('amountheld'), number of occurances ('occheld'), or whether the targets have been met for each species ('targetsmet')?}

\item{nbest}{"integer" color the n best solutions in "red".}

\item{force_reset}{"logical" should analysis be rerun even if it is stored in the cache?}
}
\value{
"prcomp" or "monoMDS" object.
}
\description{
This function makes a dendrogram to visualise differences between Marxan solutions using hierarchical clustering using various characteristics. 
Clustering can be applied to distances between raw data, or to results from dimensional reduction analysis (ie. MDS and PCA).
Distance matrices and results from multivariate analyses are cached to permit rapid display.
}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved-class}}, \code{\link{marxan}}
}

