% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{dist.MarxanResults}
\alias{dist}
\alias{dist.MarxanResults}
\alias{dist.MarxanSolved}
\title{Dissimilarity matrix for Marxan solutions}
\usage{
\method{dist}{MarxanResults}(x, var = "selections", method = "bray",
  force_reset = FALSE)

\method{dist}{MarxanSolved}(x, var = "selections", method = "bray",
  force_reset = FALSE)

dist(x, ...)
}
\arguments{
\item{x}{"MarxanResults" or "MarxanSolved" object.}

\item{var}{"character" should solutions be compared based on selections ('selections'), or the amount held ('amountheld'), number of occurances ('occheld'), or whether the targets have been met for each species ('targetsmet')?}

\item{method}{"character" name of distance metric to use for calculations (see \code{\link[vegan]{vegdist}})}

\item{force_reset}{"logical" should analysis be run even if it is stored in the cache?}
}
\value{
"dist"  with dissimilarity indices.
}
\description{
This function calculates a dissimilty matrix for Marxan solutions using various characteristics.
Results cached to permit rapid display for plotting functions.
}
\seealso{
\code{\link{MarxanSolved-class}}, \code{\link{marxan}}, \code{\link{ordiplot}}, \code{\link{dendrogram}}
}

