% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{findMarxanExecutablePath}
\alias{findMarxanExecutablePath}
\title{Find Marxan executable suitable for computer}
\usage{
findMarxanExecutablePath()
}
\value{
"logical" Is Marxan installed?
}
\description{
This function checks the computer's specifications and sets options('marxanExecutablePath') accordingly.
Marxan executables can be downloaded from \url{http://www.uq.edu.au/marxan/marxan-software}, and installed by unzipping the files contents, and copying them into the /bin folder in this package's installation directory. 
If a suitable executable cannot be found, this function will fail and provide information.
}
\examples{
# marxan executable files should be copied to this directory
system.file("bin", package="marxan")
# look for Marxan
\donttest{
findMarxanExecutablePath()
}
# was Marxan found?
is.marxanInstalled()
}
\seealso{
\code{\link{is.marxanInstalled}}.
}

