% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R
\name{format.MarxanData}
\alias{format.MarxanData}
\title{Format Data for Marxan}
\usage{
\method{format}{MarxanData}(polygons, rasters, targets = "20\%", spf = 1,
  pu = NULL, species = NULL, puvspecies = NULL, boundary = NULL, ...,
  verbose = FALSE)
}
\arguments{
\item{polygons}{"SpatialPolygons" with planning unit data}

\item{rasters}{"RasterLayer", "RasterStack", "RasterBrick" with species distribution data}

\item{targets}{"numeric" or "character" vector with absolute or percent-based targets}

\item{pu}{"data.frame" planning unit data; with columns "id", "cost", and "status" columns (optional)}

\item{species}{"data.frame" species data; with columns "id", "target", "spf", and "name" columns (optional)}

\item{puvspecies}{"data.frame" pu vs. species data; with "species", "pu", and "amount" columns (optional)}

\item{boundary}{"data.frame" with shared boundary data; with "id1", "id2", and "boundary" columns (optional)}

\item{verbose}{"logical" Should information on data pre-processing be displayed?}
}
\value{
MarxanData object
}
\description{
This function prepares spatially explicit planning unit and species data for Marxan processing.
By default, raw data can be provided which will processed for Marxan. However, if particular Marxan input file have
already been generated, these can be supplied using the pu, species, puvspecies, or boundary arguments to avoid re-processing
these outputs.
}
\examples{
data(planningunits, species, speciesRanges)
x<-MarxanData(planningunits, speciesRanges, targets="10\%")
y<-MarxanData(planningunits, speciesRanges, species=species)
stopifnot(identical(x,y))
}
\seealso{
\code{\link{read.MarxanData}}, \code{\link{write.MarxanData}}, \code{\link{MarxanData}}, \code{\link{MarxanData-class}}
}

