% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MarxanData.R, R/MarxanResults.R
\docType{methods}
\name{is.cached}
\alias{is.cached}
\alias{is.cached,MarxanData,character-method}
\alias{is.cached,MarxanResults,character-method}
\title{Test if hash is cached in a Marxan object}
\usage{
is.cached(x, name)

\S4method{is.cached}{MarxanData,character}(x, name)

\S4method{is.cached}{MarxanResults,character}(x, name)
}
\arguments{
\item{x}{"MarxanData" or "MarxanResults" object}

\item{name}{"character" hash.}
}
\value{
"logical" Is it cached?
}
\description{
Tests if hash is cached in Marxan object.
}
\note{
caches are implemented using environments, the hash is used as the name of the object in the environment.
}
\keyword{internal}

