% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanSolved.R
\docType{methods}
\name{is.comparable,MarxanUnsolvedOrSolved,MarxanUnsolvedOrSolved-method}
\alias{is.comparable,MarxanUnsolvedOrSolved,MarxanUnsolvedOrSolved-method}
\title{describeIn is.comparable}
\usage{
\S4method{is.comparable}{MarxanUnsolvedOrSolved,MarxanUnsolvedOrSolved}(x, y)
}
\description{
describeIn is.comparable
}

