% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{is.gdalInstalled}
\alias{is.gdalInstalled}
\title{Test if GDAL is installed on computer}
\usage{
is.gdalInstalled()
}
\value{
"logical" is GDAL installed?
}
\description{
This function tests if GDAL is installed on the computer.
If not, download it here: \url{http://download.osgeo.org/gdal}.
}
\examples{
is.gdalInstalled()
}
\seealso{
\code{\link[gdalUtils]{gdal_setInstallation}}.
}

