% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{is.marxanInstalled}
\alias{is.marxanInstalled}
\title{Test if Marxan is installed on computer}
\usage{
is.marxanInstalled()
}
\value{
"logical" is it installed?
}
\description{
This function determines if Marxan is installed on the computer, and will update \code{\link[base]{options}}
}
\examples{
is.marxanInstalled()
}

