% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{log.MarxanResults}
\alias{log}
\alias{log.MarxanResults}
\alias{log.MarxanSolved}
\title{Names}
\usage{
\method{log}{MarxanResults}(x)

\method{log}{MarxanSolved}(x)

log(x, ...)
}
\arguments{
\item{x}{any "MarxanOpts", "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.}

\item{x}{"MarxanResults" or "MarxanSolved" object.}
}
\description{
This function returns the species names for a Marxan object.
}
\section{Methods (by class)}{
\itemize{
\item \code{MarxanResults}: 

\item \code{MarxanSolved}: 
}}
\note{
"MaranResults" objects do not store species names.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
Log

This function returns the log file associated with running Marxan.

\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved}}, \code{\link{marxan}}
}

