% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/marxan.R
\name{marxan}
\alias{marxan}
\alias{marxan.SpatialPolygons}
\alias{marxan.character}
\alias{marxan.data.frame}
\title{General Marxan Function}
\usage{
marxan(x, ...)

\method{marxan}{character}(path, solve = TRUE)

\method{marxan}{SpatialPolygons}(polygons, rasters, ..., solve = TRUE)

\method{marxan}{data.frame}(pu, species, puvspecies, boundary, ...,
  solve = TRUE)
}
\arguments{
\item{x}{generic argument.}

\item{...}{arguments are passed to MarxanData and MarxanOpts functions.}

\item{path}{'character' file path to Marxan input parameters file.}

\item{solve}{"logical" should the problem be solved using Marxan?}

\item{polygons}{'SpatialPolyogns' object representing planning units.}

\item{rasters}{'Raster' object with species distribution data.}

\item{pu}{"data.frame" planning unit data; with "integer" 'id', "numeric" 'cost', "integer" 'status' columns.}

\item{species}{"data.frame" with species data; with "integer" 'id', "numeric" 'target', "numeric" 'spf', and "character" 'name' columns.}

\item{puvspecies}{"data.frame" with data on species density in each planning unit, with "integer" 'species', "integer" 'pu', and "numeric" 'target' columns. This "data.frame" is sorted in order of 'pu' column.}

\item{boundary}{"data.frame" with data on the shared boundary length of planning; with "integer" 'id1', "integer" 'id2', and "numeric" 'amount' columns.}
}
\value{
"MarxanSolved"  or "MarxanUnsolved"
}
\description{
This is a general function to create Marxan objects from scratch and run the Marxan program to generate solutions.
}
\note{
See the package vignette for help.
}
\seealso{
\code{\link{MarxanOpts}}, \code{\link{MarxanData}}, \code{\link{MarxanResults}}, \code{\link{MarxanUnsolved}}
}

