% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R
\name{MarxanData}
\alias{MarxanData}
\title{Create new MarxanData object}
\usage{
MarxanData(pu, species, puvspecies, boundary, polygons = NULL,
  puvspecies_spo = NULL, skipchecks = FALSE, .cache = new.env())
}
\arguments{
\item{pu}{"data.frame" planning unit data; with columns 'id', 'cost', and 'status' columns.}

\item{species}{species "data.frame" species data; with columns 'id', 'target', 'spf', and 'name' columns.}

\item{puvspecies}{"data.frame" pu vs. species data, with 'species', 'pu', and 'amount' columns, sorted by 'pu'.}

\item{boundary}{"data.frame" with shared boundary data; with "id1", "id2", and "boundary" columns.}

\item{polygons}{"PolySet" with planning unit data or "NULL" if geoplotting capabilities are not required.}

\item{puvspecies_spo}{"data.frame" with values in puvspecies sorted by species. Defaults to NULL so this is calculated automatically from argument to puvspecies.}

\item{skipchecks}{"logical" skip data integrity checks? May improve speed for big data sets. Defaults to FALSE.}

\item{.cache}{"environment" the cache used to store objects to speed up plottnig functions. Do not set this unless you know what you're doing.}
}
\value{
MarxanData object
}
\description{
This function creates a MarxanData using pre-processed data.
}
\examples{
data(planningunits, species, speciesRanges)
x<-MarxanData(
	planningunits@data,
	speciesTargets,
 calcPuVsSpeciesData(planningunits, species),
 calcBoundaryData(planningunits)
)
}
\seealso{
\code{\link{read.MarxanData}}, \code{\link{write.MarxanData}}, \code{\link{format.MarxanData}}, \code{\link{MarxanData-class}}
}

