% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanOpts.R
\docType{class}
\name{MarxanOpts-class}
\alias{MarxanOpts-class}
\title{MarxanOpts: An S4 class to represent Marxan input parameters}
\description{
This class is used to store Marxan input parameters.
}
\details{
General parameters
}
\section{Slots}{

\describe{
\item{\code{BLM}}{"numeric" boundary length modifier (default=1000000)}

\item{\code{PROP}}{"numeric" proportion of planning units in initial reserve system (default=-1)}

\item{\code{NUMREPS}}{"integer" number of replicate runs (default=100)
Annealing parameters}

\item{\code{NUMITNS}}{"integer" mumber of iterations for annealing (default=1000000)}

\item{\code{STARTTEMP}}{"numeric" initial temperature for annealing (default=-1)}

\item{\code{COOLFAC}}{"numeric" cooling factor for annealing (default=0)}

\item{\code{NUMTEMP}}{"integer" number of temperature decreases for annealing (default=100000)
Cost threshold}

\item{\code{COSTTHRESH}}{"numeric" cost threshold (default=0)}

\item{\code{THRESHPEN1}}{"numeric" size of cost threshold penalty (default=0.14)}

\item{\code{THRESHPEN2}}{"numeric" shape of cost threshold penalty (default=0)
Program control}

\item{\code{MISSLEVEL}}{"numeric" amount of target below which it is counted as 'missing' (default=1)}

\item{\code{ITIMPTYPE}}{"integer" iterative improvement type (default=1)}

\item{\code{HEURTYPE}}{"integer" heuristic type (default=0)}

\item{\code{CLUMPTYPE}}{"integer" clumping penalty type (default=0)}

\item{\code{VERBOSITY}}{"integer" amount of output displayed on the program screen (default=3)}

\item{\code{NCORES}}{"integer "number of cores to use for processing (default=1)}
}}

