% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanOpts.R
\name{MarxanOpts}
\alias{MarxanOpts}
\title{Create "MarxanOpts" object}
\usage{
MarxanOpts(..., ignore.extra = FALSE)
}
\arguments{
\item{ignore.extra}{"logical" Should extra arguments be ignored? Defaults to \code{FALSE}.}

\item{BLM}{"numeric" Boundary length modifier. Defaults to 1000000.}

\item{PROP}{"numeric" Proportion of planning units in initial reserve system. Defaults to -1.}

\item{NUMREPS}{"integer" Number of replicate runs. Defaults to 100.}

\item{NUMITNS}{"integer" Number of iterations for annealing. Defaults to 1000000.}

\item{STARTTEMP}{"numeric" Initial temperature for annealing. Default to -1.}

\item{COOLFAC}{"numeric" Cooling factor for annealing. Defaults to 0.}

\item{NUMTEMP}{"integer" Number of temperature decreases for annealing. Defaults to 100000.}

\item{COSTTHRESH}{"numeric" Cost threshold. Defaults to 0.}

\item{THRESHPEN1}{"numeric" Size of cost threshold penalty. Defaults to 0.14.}

\item{THRESHPEN2}{"numeric" Shape of cost threshold penalty. Defaults to 0.}

\item{MISSLEVEL}{"numeric" Amount of target below which it is counted as 'missing'. Defaults to 1.}

\item{ITIMPTYPE}{"integer" Iterative improvement type. Defaults to 1.}

\item{HEURTYPE}{"integer" Heuristic type. Defaults to 0.}

\item{CLUMPTYPE}{"integer" Clumping penalty type. Defaults to 0.}

\item{VERBOSITY}{"integer" Amount of output displayed on the program screen. Defaults to 3.}

\item{NCORES}{"integer" Number of cores to use for processing. Defaults to 1.}
}
\value{
"MarxanOpts" object
}
\description{
This function creates a new MarxanOpts object.
}
\examples{
x<-MarxanOpts(NCORES=4, NUMREPS=2, NUMITNS=5)
}
\seealso{
\code{\link{MarxanOpts-class}},  \code{\link{read.MarxanOpts}}, \code{\link{write.MarxanOpts}}.
}

