% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R
\docType{class}
\name{MarxanResults-class}
\alias{MarxanResults-class}
\title{MarxanResults: An S4 class to represent Marxan results}
\description{
This class is used to store the results from a Marxan run.
}
\section{Slots}{

\describe{
\item{\code{summary}}{"data.frame" with summary information on solutions.}

\item{\code{selections}}{"matrix" with binary selections.}

\item{\code{amountheld}}{"matrix" with the amount held for each species in each solution.}

\item{\code{occheld}}{"matrix" with the number of occurrences for each species in each solution.}

\item{\code{targetsmet}}{"matrix" indicating whether the targets have been met for each species in each solution.}

\item{\code{best}}{"integer" with index of best solution.}

\item{\code{log}}{"character" with Marxan log  file.}
}}
\seealso{
\code{link{MarxanResults}}, \code{link{read.MarxanResults}},
}

