% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanSolved.R
\docType{class}
\name{MarxanSolved-class}
\alias{MarxanSolved-class}
\title{MarxanSolved: An S4 class to represent Marxan inputs and outputs}
\description{
This class is used to store Marxan input and output data in addition to input parameters.
}
\section{Slots}{

\describe{
\item{\code{opts}}{"MarxanOpts" object used to store input parameters.}

\item{\code{data}}{"MarxanData" object used to store input data.}

\item{\code{results}}{"MarxanResults" object used to store results.}
}}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}, \code{\link{MarxanResults-class}}.
}

