% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarxanSolved.R
\name{MarxanSolved}
\alias{MarxanSolved}
\title{Create new "MarxanSolved" object}
\usage{
MarxanSolved(unsolved, results)
}
\arguments{
\item{unsolved}{"MarxanUnsolved" object.}

\item{results}{"MarxanResults" object.}
}
\value{
"MarxanSolved" object.
}
\description{
This function creates a "MarxanSolved" object using a "MarxanUnsolved" object and a "MarxanResults" object.
}
\seealso{
\code{\link{MarxanSolved-class}}, \code{\link{MarxanResults-class}}.
}

