% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanUnsolved.R
\docType{class}
\name{MarxanUnsolved-class}
\alias{MarxanUnsolved-class}
\title{MarxanUnsolved: An S4 class to represent Marxan inputs}
\description{
This class is used to store Marxan input data and input parameters.
}
\section{Slots}{

\describe{
\item{\code{data}}{"MarxanData" object used to store input data.}

\item{\code{opts}}{"MarxanOpts" object used to store input parameters.}
}}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}
}

