% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanUnsolved.R
\name{MarxanUnsolved}
\alias{MarxanUnsolved}
\title{Create a new MarxanUnsolved object}
\usage{
MarxanUnsolved(opts, data)
}
\arguments{
\item{opts}{"MarxanOpts" object.}

\item{data}{"MarxanData" object.}
}
\value{
"MarxanUnsolved" object.
}
\description{
This function creates a MarxanUnsolved object using MarxanOpts and MarxanData objects.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}.
}

