% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MarxanData.R, R/MarxanUnsolved.R, R/MarxanSolved.R
\name{maxtargets}
\alias{maxtargets}
\alias{maxtargets.MarxanData}
\alias{maxtargets.MarxanSolved}
\alias{maxtargets.MarxanUnsolved}
\alias{maxtargets<-.MarxanData}
\alias{maxtargets<-.MarxanSolved}
\alias{maxtargets<-.MarxanUnsolved}
\title{Maximum Targets}
\usage{
maxtargets(x, ...)

\method{maxtargets}{MarxanData}(x)

\method{maxtargets}{MarxanData}(x) <- value

\method{maxtargets}{MarxanUnsolved}(x)

\method{maxtargets}{MarxanUnsolved}(x) <- value

\method{maxtargets}{MarxanSolved}(x)

\method{maxtargets}{MarxanSolved}(x) <- value
}
\arguments{
\item{x}{any "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.}

\item{...}{not used.}

\item{value}{"numeric" with new maximum targets.}
}
\description{
This function returns or assigns the maximum targets for species in a Marxan object. Maximum targets are necessary for setting percent based targets.
}
\note{
This generic method does not work on "MaranResults" objects because they do not store this information.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

