% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarxanResults.R
\name{merge.MarxanResults}
\alias{merge.MarxanResults}
\title{Merge Marxan results}
\usage{
merge.MarxanResults(x)
}
\arguments{
\item{x}{"list" of "MarxanResults" objects.}
}
\value{
"MarxanResults" object
}
\description{
This function merges a list of "MarxanResults" objects into a single MarxanResults object. 
It is used for collating Marxan results data when Marxan runs have been run in parallel.
}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanResults}}.
}

