% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MarxanResults.R, R/MarxanSolved.R
\name{mpm}
\alias{mpm}
\alias{mpm.MarxanResults}
\alias{mpm.MarxanSolved}
\title{Extract minimum proportion met ratio}
\usage{
mpm(x, ...)

\method{mpm}{MarxanResults}(x, y = NULL)

\method{mpm}{MarxanSolved}(x, y = NULL)
}
\arguments{
\item{x}{"MarxanResults or "MarxanSolved" object.}

\item{...}{not used.}

\item{y}{"NULL" to return all ratios, "integer" 0 to return ratios for best solution, "integer" value greater than 0 for \code{y}'th solution ratios.}
}
\value{
"matrix" or "numeric" vector with solution ratios depending on arguments.
}
\description{
Extract minimum proportion met (MPM) ratio from "MarxanResults" or "MarxanSolved" object.
}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved-class}}, \code{\link{marxan}}
}

