% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{opt}
\alias{opt}
\title{Update Marxan Input Parameters}
\usage{
opt(name, value)
}
\arguments{
\item{name}{"character" name of parameter to change.}

\item{value}{"numeric" new value.}
}
\value{
"MarxanOptsOperation" object.
}
\description{
This function is used in the formula argument of the update function to change input parameters of a "MarxanOpts", "MarxanUnsolved", or "MarxanSolved" object.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}} \code{\link{update}}, \code{\link{species}}, \code{\link{pu}}
}

