% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planningunits}
\alias{planningunits}
\title{Planning units for example reserve selection problems}
\format{An object of class \code{"SpatialPolygonsDataFrame"}.}
\usage{
data(planningunits)
}
\description{
Example planning units in Tasmania (Australia) for an example reserve selection problem.
}
\examples{
data(planningunits)
\donttest{plot(planningunits)}
}
\keyword{datasets}

