% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MarxanSolved.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,MarxanSolved,MarxanSolved-method}
\alias{plot,MarxanSolved,missing-method}
\alias{plot,MarxanSolved,numeric-method}
\title{Plot Marxan solutions}
\usage{
plot(x, y, ...)

\S4method{plot}{MarxanSolved,numeric}(x, y, basemap = "none",
  colramp = "Greens", lockedincol = "#000000FF",
  lockedoutcol = "#D7D7D7FF", alpha = ifelse(basemap == "none", 1, 0.7),
  grayscale = FALSE, force_reset = FALSE)

\S4method{plot}{MarxanSolved,missing}(x, y, basemap = "none",
  colramp = "PuBu", lockedincol = "#000000FF", lockedoutcol = "#D7D7D7FF",
  alpha = ifelse(basemap == "none", 1, 0.7), grayscale = FALSE,
  force_reset = FALSE)

\S4method{plot}{MarxanSolved,MarxanSolved}(x, y, i = NULL, j = i,
  basemap = "none", colramp = ifelse(is.null(i), "RdYlBu", "Accent"),
  xlockedincol = "#000000FF", xlockedoutcol = "#D7D7D7FF",
  ylockedincol = "#FFFFFFFF", ylockedoutcol = "#D7D7D7FF",
  alpha = ifelse(basemap == "none", 1, 0.7), grayscale = FALSE,
  force_reset = FALSE)
}
\arguments{
\item{x}{"MarxanData", "MarxanUnsolved", "MarxanSolved" object.}

\item{y}{See below for details:
\itemize{ 
    \item if "missing": function plots the selection frequency of planning units for all solutions.
\item if "integer": function plots the selection for the n'th solution, use 0 to plot the best solution.
\item if "character": set to name of species to display its density, or "all" for sum density of all species, or "rich" to get species richness in each planning unit.
\item if "MarxanSolved" plots the difference in solutions.
}}

\item{...}{note used.}

\item{basemap}{"character" name of basemap to display. Valid names are "roadmap", "mobile", "satellite", "terrain", "hybrid", "mapmaker-roadmap", "mapmaker-hybrid".}

\item{colramp}{"character" name of colour palette (see \code{\link[RColorBrewer]{brewer.pal.info}}).}

\item{lockedincol}{"character" color to plot locked in planning units.}

\item{lockedoutcol}{"character" color to plot locked out planning units.}

\item{alpha}{"numeric" alpha value}

\item{grayscale}{"logical" should basemap be gray scale?}

\item{force_reset}{"logical" ignore data in cache? Setting this as ignore will make function slower but may avoid bugs in cache system.}

\item{i}{"numeric" or NULL object. Defaults to NULL. 
\itemize{
    \item if i is "NULL": differences in selection frequencies are plotted.
    \item if i is "integer" and j is "integer", plots differences selection status for solution i in x, and solution j in y. Set i or j to 0 to refer to the best solution in x or y respectively.
}}

\item{j}{"numeric" or NULL object. See \code{i} for more information. Defaults to NULL.}

\item{xlockedincol}{"character" color to plot locked in planning units in object \code{x}.}

\item{xlockedoutcol}{"character" color to plot locked out planning units in object \code{x}.}

\item{ylockedincol}{"character" color to plot locked in planning units in object \code{y}.}

\item{ylockedoutcol}{"character" color to plot locked out planning units in object \code{y}.}
}
\description{
This function makes a geoplot displaying Marxan solutions.
}
\note{
This function will return an error if spatial polygons were not supplied during the construction of the Marxan object.
}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved-class}}, \code{\link{basemap}} \code{\link{marxan}}, \code{\link{dendrogram}}, \code{\link{dotchart}}, \code{\link{ordiplot}}, \code{\link{spplot}}
}

