% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R, R/MarxanOpts.R, R/MarxanResults.R, R/MarxanSolved.R, R/MarxanUnsolved.R, R/generics.R
\name{print.MarxanData}
\alias{print}
\alias{print.MarxanData}
\alias{print.MarxanOpts}
\alias{print.MarxanResults}
\alias{print.MarxanSolved}
\alias{print.MarxanUnsolved}
\title{Print}
\usage{
\method{print}{MarxanData}(x, header = TRUE)

\method{print}{MarxanOpts}(x, header = TRUE)

\method{print}{MarxanResults}(x, header = TRUE)

\method{print}{MarxanSolved}(x)

\method{print}{MarxanUnsolved}(x)

print(x, ...)

print(x, ...)
}
\arguments{
\item{x}{any "MarxanOpts", "MarxanData", "MarxanResults", "MarxanUnsolved", or "MarxanSolved" object.}

\item{x}{any "MarxanOpts", "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.}
}
\description{
This function returns information about a Marxan object.

This function returns the species names for a Marxan object.
}
\section{Methods (by class)}{
\itemize{
\item \code{MarxanData}: 

\item \code{MarxanOpts}: 

\item \code{MarxanResults}: 

\item \code{MarxanSolved}: 

\item \code{MarxanUnsolved}: 
}}
\note{
"MaranResults" objects do not store species names.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}, \code{\link{MarxanResults-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}},

\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

