% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{pu}
\alias{pu}
\title{Update Marxan Planning Unit Parameters}
\usage{
pu(id, cost = NA, status = NA)
}
\arguments{
\item{id}{"integer" id of the planning unit.}

\item{cost}{"numeric" new cost value.}

\item{status}{"numeric" new status value.}
}
\value{
"MarxanPuOperation" object.
}
\description{
This function is used in the formula argument of the update function to change planning unit parameters of a "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.
}
\note{
Set argument 'cost' or 'status' to NA (default) to keep the same.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}} \code{\link{update}}, \code{\link{opt}}, \code{\link{species}}
}

