% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R, R/MarxanSolved.R, R/MarxanUnsolved.R, R/generics.R
\name{puids.MarxanData}
\alias{puids}
\alias{puids.MarxanData}
\alias{puids.MarxanSolved}
\alias{puids.MarxanUnsolved}
\alias{puids<-.MarxanData}
\alias{puids<-.MarxanSolved}
\alias{puids<-.MarxanUnsolved}
\title{Planning unit identifiers}
\usage{
\method{puids}{MarxanData}(x)

\method{puids}{MarxanData}(x) <- value

\method{puids}{MarxanSolved}(x)

\method{puids}{MarxanSolved}(x) <- value

\method{puids}{MarxanUnsolved}(x)

\method{puids}{MarxanUnsolved}(x) <- value

puids(x, ...)
}
\arguments{
\item{x}{any "MarxanOpts", "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.}
}
\description{
This function returns or assigns the planning unit ids for a Marxan object.
}
\section{Methods (by class)}{
\itemize{
\item \code{MarxanData}: 

\item \code{MarxanData}: 

\item \code{MarxanSolved}: 

\item \code{MarxanSolved}: 

\item \code{MarxanUnsolved}: 

\item \code{MarxanUnsolved}: 
}}
\note{
This generic method does not work on "MaranResults" objects because they do not store this information.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

