% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{methods}
\name{rasterize.gdal}
\alias{rasterize.gdal}
\alias{rasterize.gdal,SpatialPolygonsDataFrame,RasterLayer-method}
\title{Rasterize polygon data using GDAL}
\usage{
rasterize.gdal(x, y, ...)

\S4method{rasterize.gdal}{SpatialPolygonsDataFrame,RasterLayer}(x, y,
  field = NULL)
}
\arguments{
\item{x}{"SpatialPolygonsDataFrame" object.}

\item{y}{"RasterLayer" with dimensions, extent, and resolution to be used as a template for new raster.}

\item{...}{not used.}

\item{field}{"character" column name with values to burn into the output raster. If not supplied, default behaviour is to burn polygon indices into the "RasterLayer".}
}
\value{
"RasterLayer" object.
}
\description{
This function converts a "SpatialPolygonsDataFrame" to a "RasterLayer" using GDAL.
It is expected to be faster than \code{\link[raster]{rasterize}} for large datasets.
However, it will be significantly slower for small datasets because the data will need to be written and read from disk.
}
\examples{
data(taspu,tasinvis)
x<-rasterize.gdal(taspu[1:5,],tasinvis[[1]])
}
\seealso{
\code{\link[raster]{rasterize}}, \code{\link{is.gdalInstalled}}.
}

