% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R
\name{read.MarxanData}
\alias{read.MarxanData}
\title{Read Input Marxan Data from Disk}
\usage{
read.MarxanData(path, ...)
}
\arguments{
\item{path}{"character" file path for input parameter file}
}
\value{
MarxanData object
}
\description{
This function loads Marxan input data from disk.
}
\seealso{
\code{\link{write.MarxanData}}, \code{\link{format.MarxanData}}, \code{\link{MarxanData}}, \code{\link{MarxanData-class}}
}

