% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarxanOpts.R
\name{read.MarxanOpts}
\alias{read.MarxanOpts}
\title{Read Marxan input parameters from disk}
\usage{
read.MarxanOpts(path)
}
\arguments{
\item{path}{"character" directory path for location to save input parameters file.}
}
\description{
This function reads Marxan parameter settings from an input file.
}
\examples{
x<-MarxanOpts()
write.MarxanData(x, file.path(tempdir(), 'input.dat'))
y<-read.MarxanData(file.path(tempdir(), 'input.dat'))
stopifnot(identical(x,y))
}
\seealso{
\code{\link{MarxanOpts-class}},  \code{\link{MarxanOpts}}, \code{\link{write.MarxanOpts}}.
}

