% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R
\name{read.MarxanResults}
\alias{read.MarxanResults}
\title{Read Marxan results from disk}
\usage{
read.MarxanResults(dir)
}
\arguments{
\item{dir}{"character" with file path to directory with Marxan output files.}
}
\value{
"MarxanResults" object
}
\description{
This function reads outputs from a directory after Marxan been run.
}
\seealso{
\code{link{MarxanResults-class}}, \code{link{MarxanResults}}
}

