% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanSolved.R
\name{read.MarxanSolved}
\alias{read.MarxanSolved}
\title{Read Marxan input data, parameters and results}
\usage{
read.MarxanSolved(path, skipchecks = FALSE)
}
\arguments{
\item{path}{"character" path for input file to load.}

\item{skipchecks}{"logical" Should data integrity checks be skipped?}
}
\description{
This function saves "MarxanSolved" objects to disk.
}
\seealso{
\code{\link{MarxanSolved}}, \code{\link{MarxanSolved-class}}, \code{\link{read.MarxanOpts}}, \code{\link{read.MarxanData}}, \code{\link{read.MarxanUnsolved}}.
}

