% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MarxanUnsolved.R
\name{read.MarxanUnsolved}
\alias{read.MarxanUnsolved}
\title{Read Marxan input data and parameters}
\usage{
read.MarxanUnsolved(path, skipchecks = FALSE)
}
\arguments{
\item{path}{"character" path for input file to load.}

\item{skipchecks}{"logical" should data integrity checks be skipped?}
}
\description{
This function saves "MarxanUnsolved" objects to disk.
}
\seealso{
\code{\link{MarxanUnsolved}}, \code{\link{MarxanUnsolved-class}}, \code{\link{read.MarxanOpts}}, \code{\link{read.MarxanData}}.
}

