% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{score.MarxanResults}
\alias{score}
\alias{score.MarxanResults}
\alias{score.MarxanSolved}
\title{Extract solution score}
\usage{
\method{score}{MarxanResults}(x, y = NULL)

\method{score}{MarxanSolved}(x, y = NULL)

score(x, ...)
}
\arguments{
\item{x}{"MarxanResults or "MarxanSolved" object.}

\item{y}{"NULL" to return all scores, "integer" 0 to return score for best solution, "integer" value greater than 0 for \code{y}'th solution score.}
}
\value{
"matrix" or "numeric" vector with solution score(s) depending on arguments.
}
\description{
Extract solution score from "MarxanResults" or "MarxanSolved" object.
}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved-class}}, \code{\link{marxan}}
}

