% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{selections.MarxanResults}
\alias{selections}
\alias{selections.MarxanResults}
\alias{selections.MarxanSolved}
\title{Extract solution selections}
\usage{
\method{selections}{MarxanResults}(x, y = NULL)

\method{selections}{MarxanSolved}(x, y = NULL)

selections(x, ...)
}
\arguments{
\item{x}{"MarxanResults or "MarxanSolved" object.}

\item{y}{"NULL" to return all solution selections, "integer" 0 to return selection for best solution, "integer" value greater than 0 for \code{y}'th solution selection.}
}
\value{
"matrix" or "numeric" vector with planning units statuses depending on arguments.
}
\description{
Extract selections for a given solution from a "MarxanResults or "MarxanSolved" object.
}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved-class}}, \code{\link{marxan}}.
}

