% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{sepacheived.MarxanResults}
\alias{sepacheived}
\alias{sepacheived.MarxanResults}
\alias{sepacheived.MarxanSolved}
\title{Extract separation achieved value}
\usage{
\method{sepacheived}{MarxanResults}(x, y = NULL)

\method{sepacheived}{MarxanSolved}(x, y = NULL)

sepacheived(x, ...)
}
\arguments{
\item{x}{"MarxanResults or "MarxanSolved" object.}

\item{y}{"missing" to return all values, "integer" 0 to return values for best solution, "integer" value greater than 0 for y'th solution values.}
}
\value{
"matrix" or "numeric" vector with values depending on arguments.
}
\description{
Extract separation achieved value from "MarxanResults" or "MarxanSolved" object.
}
\section{Methods (by class)}{
\itemize{
\item \code{MarxanResults}: 

\item \code{MarxanSolved}: 
}}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved-class}}, \code{\link{marxan}}
}

