% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Polygons2PolySet.R
\docType{methods}
\name{SpatialPolygons2PolySet,SpatialPolygons-method}
\alias{SpatialPolygons2PolySet,SpatialPolygons-method}
\title{describeIn SpatialPolygons2PolySet}
\usage{
\S4method{SpatialPolygons2PolySet}{SpatialPolygons}(x, n_preallocate = 10000L)
}
\description{
describeIn SpatialPolygons2PolySet
}

