% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Polygons2PolySet.R
\docType{methods}
\name{SpatialPolygons2PolySet,SpatialPolygonsDataFrame-method}
\alias{SpatialPolygons2PolySet,SpatialPolygonsDataFrame-method}
\title{describeIn SpatialPolygons2PolySet}
\usage{
\S4method{SpatialPolygons2PolySet}{SpatialPolygonsDataFrame}(x,
  n_preallocate = 10000L)
}
\description{
describeIn SpatialPolygons2PolySet
}

