% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{Simulated species data}
\format{An object of class \code{"RasterStack"}.}
\usage{
data(species)
}
\description{
Data for five simulated species inhabiting Tasmania (Australia) for an example reserve selection problem.
}
\examples{
data(species)
\donttest{plot(species)}
}
\keyword{datasets}

