% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MarxanData.R, R/MarxanUnsolved.R, R/MarxanSolved.R
\name{spfs}
\alias{spfs}
\alias{spfs.MarxanData}
\alias{spfs.MarxanSolved}
\alias{spfs.MarxanUnsolved}
\alias{spfs<-}
\alias{spfs<-.MarxanData}
\alias{spfs<-.MarxanSolved}
\alias{spfs<-.MarxanUnsolved}
\title{Species penalty factors}
\usage{
spfs(x, ...)

spfs(x, ...) <- value

\method{spfs}{MarxanData}(x)

\method{spfs}{MarxanData}(x) <- value

\method{spfs}{MarxanUnsolved}(x)

\method{spfs}{MarxanUnsolved}(x) <- value

\method{spfs}{MarxanSolved}(x)

\method{spfs}{MarxanSolved}(x) <- value
}
\arguments{
\item{x}{any "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.}

\item{...}{not used.}

\item{value}{"numeric" new species penalty factors.}
}
\description{
This function returns or assigns the species penalty factors for speces in a Marxan object.
}
\note{
This generic method does not work on "MaranResults" objects because they do not store this information.
}
\seealso{
\code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

