% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarxanData.R
\name{spp}
\alias{spp}
\title{Update Marxan species parameters}
\usage{
spp(x, name = NA, spf = NA, target = NA)
}
\arguments{
\item{x}{"numeric" species id or "character" species name.}

\item{name}{"character" new species name.}

\item{spf}{"numeric" new species penalty factor.}

\item{target}{"numeric" new target.}
}
\value{
"MarxanSpeciesOperation" object.
}
\description{
This function is used in the formula argument of the update function to change species parameters of a "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.
}
\note{
Only specified parameters are changed, all unspecific parameters will remain unchanged.
}
\examples{
spp(1, name='species1')
spp(2, spf=5)
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}} \code{\link{update}}, \code{\link{opt}}, \code{\link{pu}}.
}

