% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{summary.MarxanResults}
\alias{summary}
\alias{summary.MarxanResults}
\alias{summary.MarxanSolved}
\title{Summary}
\usage{
\method{summary}{MarxanResults}(x)

\method{summary}{MarxanSolved}(x)

summary(x, ...)
}
\arguments{
\item{x}{"MarxanSolved" or "MarxanResults" object.}
}
\value{
"data.frame" with solution information.
}
\description{
This function returns the summary table output by Marxan.
}
\section{Methods (by class)}{
\itemize{
\item \code{MarxanResults}: 

\item \code{MarxanSolved}: 
}}
\seealso{
\code{\link{MarxanSolved-class}}, \code{\link{MarxanResults}}, \code{\link{marxan}}
}

