% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MarxanData.R, R/MarxanSolved.R, R/MarxanUnsolved.R, R/generics.R
\name{targets.MarxanData}
\alias{targets}
\alias{targets.MarxanData}
\alias{targets.MarxanSolved}
\alias{targets.MarxanUnsolved}
\alias{targets<-}
\alias{targets<-.MarxanData}
\alias{targets<-.MarxanSolved}
\alias{targets<-.MarxanUnsolved}
\title{Targets}
\usage{
\method{targets}{MarxanData}(x)

\method{targets}{MarxanData}(x) <- value

\method{targets}{MarxanSolved}(x)

\method{targets}{MarxanSolved}(x) <- value

\method{targets}{MarxanUnsolved}(x)

\method{targets}{MarxanUnsolved}(x) <- value

targets(x, ...)

targets(x, ...) <- value
}
\arguments{
\item{x}{any "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.}

\item{...}{not used.}
}
\description{
This function returns or assigns the targets for species in a Marxan object.
}
\note{
This generic method does not work on "MaranResults" objects because they do not store this information.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

