% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tasinvis}
\alias{tasinvis}
\title{Broad vegetation data}
\format{An object of class \code{"RasterLayer"}.}
\usage{
data(tasinvis)
}
\description{
"RasterLayer" data for the distribution of 63 vegetation classes in Tasmania, Australia. This data was sourced from the "Introduction to Marxan" course.
}
\examples{
data(tasinvis)
plot(tasinvis)
}
\seealso{
\code{\link{taspu}}
}
\keyword{datasets}

