% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taspu}
\alias{taspu}
\title{Planning units for example reserve selection problem}
\format{An object of class \code{"SpatialPolygonsDataFrame"}.}
\usage{
data(taspu)
}
\description{
"SpatialPolygonsDataFrame" data for planning units in Tasmania, Australia. This data was sourced from the "Introduction to Marxan" course. The 'data' slot contains a 'data.frame' with three fields:
\itemize{
	\item The 'id' field contains "integer" planning unit ids.
	\item The 'cost' field contains "numeric" unimproved land values to represent acquisition costs.
	\item The 'status' field indicates the level of existing protection in each planning unit. Units with 50\% or more of their area contained in IUCN protected areas are associated with a value of 2; otherwise they are associated with a value of 0.
}
}
\examples{
data(taspu)
plot(taspu)
}
\seealso{
\code{\link{tasinvis}}
}
\keyword{datasets}

