% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R
\name{write.MarxanData}
\alias{write.MarxanData}
\title{Write Marxan input data}
\usage{
write.MarxanData(x, dir = getwd(), ...)
}
\arguments{
\item{x}{"MarxanData" object to save.}

\item{dir}{"character" directory path for location to save data.}
}
\description{
This function saves "MarxanData" objects to disk.
}
\seealso{
\code{\link{read.MarxanData}}, \code{\link{format.MarxanData}}, \code{\link{MarxanData}}, \code{\link{MarxanData-class}}.
}

