% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanSolved.R
\name{write.MarxanSolved}
\alias{write.MarxanSolved}
\title{Write Marxan input data, parameters and results}
\usage{
write.MarxanSolved(x, dir = getwd())
}
\arguments{
\item{x}{"MarxanSolved" object to save.}

\item{dir}{"character" directory path for location to save data.}
}
\description{
This function saves "MarxanSolved" objects to disk.
}
\seealso{
\code{\link{MarxanSolved}}, \code{\link{MarxanSolved-class}}, \code{\link{write.MarxanOpts}}, \code{\link{write.MarxanData}}, \code{\link{write.MarxanUnsolved}}.
}

