% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanUnsolved.R
\name{write.MarxanUnsolved}
\alias{write.MarxanUnsolved}
\title{Write Marxan input data and parameters}
\usage{
write.MarxanUnsolved(x, dir = getwd())
}
\arguments{
\item{x}{"MarxanUnsolved" object to save.}

\item{dir}{"character" directory path for location to save data.}
}
\description{
This function saves "MarxanUnsolved" objects to disk.
}
\seealso{
\code{\link{MarxanUnsolved}}, \code{\link{MarxanUnsolved-class}}, \code{\link{write.MarxanOpts}}, \code{\link{write.MarxanData}}.
}

