% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zonalSum.R
\docType{methods}
\name{zonalSum}
\alias{zonalSum}
\alias{zonalSum,RasterLayer,RasterLayer-method}
\alias{zonalSum,RasterLayer,RasterStackOrBrick-method}
\title{Zonal Sum}
\usage{
zonalSum(x, y, ...)

\S4method{zonalSum}{RasterLayer,RasterStackOrBrick}(x, y,
  speciesNames = names(y), ncores = 1)

\S4method{zonalSum}{RasterLayer,RasterLayer}(x, y, speciesNames = names(y),
  ncores = 1)
}
\arguments{
\item{x}{SpatialPolygons or SpatialPolygons object}

\item{y}{RasterLayer, RasterStack, or RasterBrick}

\item{ncores}{Number of cores to use for processing data.}

\item{speciesName}{character vector of species names}
}
\value{
data.frame with sum of raster values in each polygon.
}
\description{
This function calculates the zonal sum of SpatialPolygons in a RasterLayer, RasterStack, or RasterBrick.
For most cases it is expected to run faster than the zonal functon in the raster package.
This function is useful for extracting the sum of raster values inside each planning unit.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = RasterLayer,y = RasterStackOrBrick}: 

\item \code{x = RasterLayer,y = RasterLayer}: 
}}
\examples{
data(species, planningunits)
purast<-rasterize(planningunits, species[[1]])
zonalSum(purast, species[[1]])
zonalSum(purast, species)
}
\seealso{
\code{\link[raster]{zonal}}
}

