% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{amountheld.MarxanResults}
\alias{amountheld}
\alias{amountheld.MarxanResults}
\alias{amountheld.MarxanSolved}
\title{Amount Held}
\usage{
\method{amountheld}{MarxanResults}(x, y = NULL)

\method{amountheld}{MarxanSolved}(x, y = NULL)

amountheld(x, ...)
}
\arguments{
\item{x}{"MarxanResults" or "MarxanSolved" object.}
}
\value{
"matrix" or "numeric" vector depending on arguments.
}
\description{
This function returns the amount held for each species in a solution.
}
\section{Methods (by class)}{
\itemize{
\item \code{MarxanResults}: 

\item \code{MarxanSolved}: 
}}
\seealso{
\code{link{MarxanResults-class}}, \code{link{MarxanSolved}}, \code{\link{marxan}}
}

