% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R, R/MarxanSolved.R, R/MarxanUnsolved.R, R/generics.R
\name{basemap.MarxanData}
\alias{basemap}
\alias{basemap.MarxanData}
\alias{basemap.MarxanSolved}
\alias{basemap.MarxanUnsolved}
\title{Basemap}
\usage{
\method{basemap}{MarxanData}(x, basemap = "hybrid", grayscale = FALSE,
  force_reset = FALSE)

\method{basemap}{MarxanSolved}(x, basemap = "none", grayscale = FALSE,
  force_reset = FALSE)

\method{basemap}{MarxanUnsolved}(x, basemap = "none", grayscale = FALSE,
  force_reset = FALSE)

basemap(x, ...)
}
\arguments{
\item{basemap}{"character" name of base map to display. Valid names are "roadmap", "mobile", "satellite", "terrain", "hybrid", "mapmaker-roadmap", "mapmaker-hybrid".}

\item{grayscale}{"logical" should base map be gray scale?}

\item{force_reset}{"logical" ignore data in cache? Setting this as ignore will make function slower but may avoid bugs in cache system.}

\item{xzoom}{"numeric" zoom the geoplot in or out along the x-axis by this percent.}

\item{yzoom}{"numeric" zoom the geoplot in or out along the y-axis by this percent.}
}
\value{
"list" with google map data.
}
\description{
This function retrieves google map data for planning units. The google map data is cached to provide fast plotting capabilities.
}
\section{Methods (by class)}{
\itemize{
\item \code{MarxanData}: 

\item \code{MarxanSolved}: 

\item \code{MarxanUnsolved}: 
}}
\seealso{
\cite{\link[RgoogleMaps]{GetMap.bbox}}, \cite{\link{plot}}
}

