% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MarxanData.R, R/MarxanResults.R
\docType{methods}
\name{cache}
\alias{cache}
\alias{cache,MarxanData,character,ANY-method}
\alias{cache,MarxanData,character,missing-method}
\alias{cache,MarxanResults,character,ANY-method}
\alias{cache,MarxanResults,character,missing-method}
\title{Get and Set Cache Methods}
\usage{
cache(x, name, y)

\S4method{cache}{MarxanData,character,ANY}(x, name, y)

\S4method{cache}{MarxanData,character,missing}(x, name, y)

\S4method{cache}{MarxanResults,character,ANY}(x, name, y)

\S4method{cache}{MarxanResults,character,missing}(x, name, y)
}
\arguments{
\item{x}{"MarxanData" or "MarxanResults" object}

\item{name}{"character" hash.}

\item{y}{if "ANY" this object gets cached with name, else if "missing" the object hashed at name gets returned.}
}
\value{
"any" or "NULL" depends on y argument.
}
\description{
Getter and setter methods for caches in MarxanData and MarxanResults object.
}
\note{
caches are implemented using environments, the hash is used as the name of the object in the environment.
}
\keyword{internal}

