% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcBoundaryData.R
\name{calcBoundaryData}
\alias{calcBoundaryData}
\title{Calculate Boundary Data}
\usage{
calcBoundaryData(x, ...)
}
\arguments{
\item{x}{PolyData, SpatialPolygons or SpatialPolyognsDataFrame.}

\item{tolerance}{numeric to specify precision of vertices.}

\item{lengthFactor}{numeric to scale boundary lengths.}

\item{edgeFactor}{numeric to scale boundary lengths that do not have any neighbors.}
}
\value{
data.frame with columns 'id1', 'id2', and 'amount'.
}
\description{
This function calculates boundary length data for PolyData, SpatialPolygons, and SpatialPolygonsDataFrame objects.
Be aware that this function is designed to be as fast as possible, as a result it depends on C++ code and if this
function is used improperly, then it will crash R. Furthermore, multipart polygons with touching edges will likely result in inaccuracies.
If argument set to SpatialPolygons or SpatialPolygonsDataFrame, this will be converted to PolyData before processing.
}
\examples{
data(planningunits)
bound.dat <- calcBoundaryData(planningunits)
summary(bound.dat)
}
\seealso{
this function is based on the QMARXAN algorithm \code{\url{http://aproposinfosystems.com/products/qmarxan/}} for calculating boundary length.
}

