% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcPuVsSpeciesData.R
\name{calcPuVsSpeciesData}
\alias{calcPuVsSpeciesData}
\title{Calculate Planning Units vs. Species Data}
\usage{
calcPuVsSpeciesData(x, y, speciesNames = names(y), ncores = 1,
  gdal = FALSE, field = NULL)
}
\arguments{
\item{x}{SpatialPolygons or SpatialPolygonsDataFrame object}

\item{y}{RasterLayer, RasterStack, or RasterBrick}

\item{speciesNames}{character vector of species names}

\item{ncores}{Number of cores to use for processing}

\item{gdal}{Logical. Should raster processing be performed using python gdal libraries?}

\item{field}{character if not NULL and \code{x} is a SpatialPolygonsDataFrame object, the index or name of the column with planning unit ids}
}
\value{
data.frame with sum of raster values in each polygon.
}
\description{
This function calculates the sum of species values in each planning unit.
Be aware that using polygons with overlaps will result in inaccuracies.
By default all polygons will be treated as having seperate ids.
}
\examples{
data(species, planningunits)
calcPuVsSpeciesData(planningunits, species[[1]])
calcPuVsSpeciesData(planningunits, species)
}
\seealso{
\code{\link{is.gdalInstalled}}, \code{\link{zonalSum}}
}

