% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MarxanData.R, R/MarxanUnsolved.R, R/MarxanSolved.R
\name{costs}
\alias{costs}
\alias{costs.MarxanData}
\alias{costs.MarxanSolved}
\alias{costs.MarxanUnsolved}
\alias{costs<-}
\alias{costs<-.MarxanData}
\alias{costs<-.MarxanSolved}
\alias{costs<-.MarxanUnsolved}
\title{Planning unit costs}
\usage{
costs(x, ...)

costs(x, ...) <- value

\method{costs}{MarxanData}(x)

\method{costs}{MarxanData}(x) <- value

\method{costs}{MarxanUnsolved}(x)

\method{costs}{MarxanUnsolved}(x) <- value

\method{costs}{MarxanSolved}(x)

\method{costs}{MarxanSolved}(x) <- value
}
\arguments{
\item{x}{any "MarxanOpts", "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.}

\item{...}{not used}

\item{value}{"numeric" value of new costs.}
}
\description{
This function returns or assigns the planing unit costs for a Marxan object.
}
\note{
This generic method does not work on "MaranResults" objects because they do not store this information.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

