% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{dotchart.MarxanResults}
\alias{dotchart}
\alias{dotchart.MarxanResults}
\alias{dotchart.MarxanSolved}
\title{Dot chart of Marxan solutions}
\usage{
\method{dotchart}{MarxanResults}(x, var = "score", nbest = 1, n = 50)

\method{dotchart}{MarxanSolved}(x, var = "score", nbest = 1, n = 50)

dotchart(x, ...)
}
\arguments{
\item{x}{"MarxanResults" or "MarxanSolved".}

\item{var}{"character" What variable should be used to compare solutions?}

\item{nbest}{"integer" Color the n best solutions in "red".}

\item{n}{"integer" Number of solutions to plot. Defaults to 50.}
}
\description{
This function makes a dot chart to visualise differences between Marxan solutions using summary variables. Size of dots indicate overall solution quality (ie. 'score').
}
\note{
Supported variables are shown below.
\tabular{ccl}{
	Short name \tab Full name \tab Description\cr
	'score' \tab 'Score' \tab Quality of solution.\cr
	'cost' \tab 'Cost' \tab Total cost of solution.\cr
	'npu' \tab 'Planning_Units' \tab Number of planning units selected in prioritisation.\cr
	'con' \tab 'Connectivity' \tab Sum boundary length of all edges in selected planning units' that have neighbours.\cr
	'confrac' \tab 'Connectivity_In_Fraction' \tab Connectivity efficiency relative to total boundary length.\cr
	'conin' \tab 'Connectivity_In' \tab Sum boundary length of edges belonging to selected planning units' that have neighbours.\cr
	'conout' \tab 'Connectivity_Out' \tab Sum boundary length of edges belonging to selected planning units' that have neighbours.\cr
 'penalty' \tab 'Penalty' \tab Total species penalty.\cr
 'shortfall' \tab 'Shortfall' \tab Total shortfall for species targets.\cr
 'mv' \tab 'Missing_Values' \tab Number of species that do not have their targets met.\cr
}
}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved-class}}, \code{\link{marxan}}
}

