% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{dotchart.MarxanResults}
\alias{dotchart}
\alias{dotchart.MarxanResults}
\alias{dotchart.MarxanSolved}
\title{Dot Chart of Marxan Solutions}
\usage{
\method{dotchart}{MarxanResults}(x, var = "score", nbest = 1)

\method{dotchart}{MarxanSolved}(x, var = "score", nbest = 1)

dotchart(x, ...)
}
\arguments{
\item{x}{"MarxanResults" or "MarxanSolved".}

\item{var}{"character" what variable should be used to compare solutions?}

\item{nbest}{"integer" color the n best solutions in "red".}
}
\description{
This function makes a dot chart to visualise differences between Marxan solutions using summary variables. Size of dots indicate overall solution quality (ie. 'score').
Supported summary variables are:
\tabular{rrr} {
	short name \tab full name \tab Description.
	'score' \tab 'Score' \tab quality of solution.
	'cost' \tab 'Cost' \tab total cost of solution.
	'npu' \tab 'Planning_Units' \tab number of planning units selected in prioritisation.
	'con' \tab 'Connectivity' \tab sum boundary length of all edges in selected planning units' that have neighbours.
	'confrac' \tab 'Connectivity_In_Fraction' \tab connectivity efficiency relative to total boundary length.
	'conin' \tab 'Connectivity_In' \tab sum boundary length of edges belonging to selected planning units' that have neighbours.
	'conout' \tab 'Connectivity_Out' \tab sum boundary length of edges belonging to selected planning units' that have neighbours.
 'penalty' \tab 'Penalty' \tab total species penalty .
 'shortfall' \tab 'Shortfall' \tab total shortfall for species targets .
 'mv' \tab 'Missing_Values' \tab number of species that do not have their targets met.
}
}
\section{Methods (by class)}{
\itemize{
\item \code{MarxanResults}: 

\item \code{MarxanSolved}: 
}}
\seealso{
\code{link{MarxanResults-class}}, \code{link{MarxanSolved-class}}, \code{link{marxan}}
}

