% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{findMarxanExecutablePath}
\alias{findMarxanExecutablePath}
\title{Find Marxan executable suitable for computer}
\usage{
findMarxanExecutablePath()
}
\value{
Logical. Is it installed?
}
\description{
This function checks the computer's specifications and sets options('marxanExecutablePath') accordingly.
Marxan exectuables can be downloaded from \code{\link{marxanURL}}, and installed by unzipping the files contents, and copying them
into the /bin folder in this package's installation directory.
If a suitable executable cannot be found, this function will fail and provide information. determines if Marxan is installed on the computer, and will update \code{\link[base]{options}}
}
\examples{
# Marxan executable files should be copied to this directory
system.file("bin", package="marxan")
# Look for Marxan
\donttest{
findMarxanExecutablePath()
}
# Was Marxan found?
is.marxanInstalled()
}
\seealso{
\link{\code{is.marxanInstalled}}
}

