% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MarxanData.R, R/MarxanSolved.R, R/MarxanUnsolved.R, R/generics.R
\name{inistatus.MarxanData}
\alias{inistatus}
\alias{inistatus.MarxanData}
\alias{inistatus.MarxanSolved}
\alias{inistatus.MarxanUnsolved}
\alias{inistatus<-}
\alias{inistatus<-.MarxanData}
\alias{inistatus<-.MarxanSolved}
\alias{inistatus<-.MarxanUnsolved}
\title{Planning unit initial status}
\usage{
\method{inistatus}{MarxanData}(x)

\method{inistatus}{MarxanData}(x) <- value

\method{inistatus}{MarxanSolved}(x)

\method{inistatus}{MarxanSolved}(x) <- value

\method{inistatus}{MarxanUnsolved}(x)

\method{inistatus}{MarxanUnsolved}(x) <- value

inistatus(x, ...)

inistatus(object) <- value
}
\arguments{
\item{x}{any "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.}
}
\description{
This function returns or assigns the planing unit initial statuses for a Marxan object.
}
\note{
This generic method does not work on "MaranOpts" or "MaranResults" objects because they do not store this information.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

