% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R, R/MarxanSolved.R, R/MarxanUnsolved.R, R/generics.R
\name{inistatus.MarxanData}
\alias{inistatus}
\alias{inistatus.MarxanData}
\alias{inistatus.MarxanSolved}
\alias{inistatus.MarxanUnsolved}
\alias{inistatus<-.MarxanData}
\alias{inistatus<-.MarxanSolved}
\alias{inistatus<-.MarxanUnsolved}
\title{Planning unit initial status}
\usage{
\method{inistatus}{MarxanData}(x)

\method{inistatus}{MarxanData}(x) <- value

\method{inistatus}{MarxanSolved}(x)

\method{inistatus}{MarxanSolved}(x) <- value

\method{inistatus}{MarxanUnsolved}(x)

\method{inistatus}{MarxanUnsolved}(x) <- value

inistatus(x, ...)
}
\arguments{
\item{x}{any "MarxanOpts", "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.}
}
\description{
This function returns or assigns the planing unit initial statuses for a Marxan object.
}
\section{Methods (by class)}{
\itemize{
\item \code{MarxanData}: 

\item \code{MarxanData}: 

\item \code{MarxanSolved}: 

\item \code{MarxanSolved}: 

\item \code{MarxanUnsolved}: 

\item \code{MarxanUnsolved}: 
}}
\note{
This generic method does not work on "MaranResults" objects because they do not store this information.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

