% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R, R/MarxanResults.R, R/generics.R
\docType{methods}
\name{is.cached,MarxanData,character-method}
\alias{is.cached}
\alias{is.cached,MarxanData,character-method}
\alias{is.cached,MarxanResults,character-method}
\title{Test if hash is cached in a Marxan object}
\usage{
\S4method{is.cached}{MarxanData,character}(x, name)

\S4method{is.cached}{MarxanResults,character}(x, name)

is.cached(x, name)

is.cached(x, name)
}
\arguments{
\item{x}{"MarxanData" or "MarxanResults" object}

\item{name}{"character" hash.}

\item{x}{"MarxanData" or "MarxanResults" object}

\item{name}{"character" hash.}
}
\value{
"logical" Is it cached?

"logical" Is it cached?
}
\description{
Tests if hash is cached in Marxan object.

Tests if hash is cached in Marxan object. Hashes are generated using function calls.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = MarxanData,name = character}: 

\item \code{x = MarxanResults,name = character}: 
}}
\note{
caches are implemented using environments, the hash is used as the name of the object in the environment.

caches are implemented using environments, the hash is used as the name of the object in the environment.
}
\keyword{internal}

