% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MarxanData.R, R/MarxanSolved.R
\docType{methods}
\name{is.comparable}
\alias{is.comparable}
\alias{is.comparable,MarxanData,MarxanData-method}
\alias{is.comparable,MarxanData,MarxanUnsolvedOrSolved-method}
\alias{is.comparable,MarxanUnsolvedOrSolved,MarxanData-method}
\alias{is.comparable,MarxanUnsolvedOrSolved,MarxanUnsolvedOrSolved-method}
\title{Compare Marxan objects}
\usage{
is.comparable(x, y)

\S4method{is.comparable}{MarxanData,MarxanData}(x, y)

\S4method{is.comparable}{MarxanUnsolvedOrSolved,MarxanUnsolvedOrSolved}(x, y)

\S4method{is.comparable}{MarxanData,MarxanUnsolvedOrSolved}(x, y)

\S4method{is.comparable}{MarxanUnsolvedOrSolved,MarxanData}(x, y)
}
\arguments{
\item{x}{"MarxanData", "MarxanUnsolved", or "MarxanSolved" objects.}

\item{y}{"MarxanData", "MarxanUnsolved", or "MarxanSolved" objects.}
}
\value{
"logical" are the objects based on the same data?
}
\description{
This function checks objects to see if they share the same input data.
}
\seealso{
\code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

