% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R, R/generics.R
\docType{methods}
\name{is.comparable,MarxanData,MarxanData-method}
\alias{is.comparable}
\alias{is.comparable,MarxanData,MarxanData-method}
\title{Compare Marxan objects}
\usage{
\S4method{is.comparable}{MarxanData,MarxanData}(x, y)

is.comparable(x, y)
}
\arguments{
\item{x}{"MarxanData", "MarxanUnsolved", or "MarxanSolved" objects.}
}
\value{
"logical" are the objects based on the same data?
}
\description{
This function checks objects to see if they share the same input data.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = MarxanData,y = MarxanData}: export
}}
\seealso{
\code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

