% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{is.comparable}
\alias{is.comparable}
\title{Compare Marxan Objects}
\usage{
is.comparable(x, y)
}
\arguments{
\item{x}{"MarxanData", "MarxanUnsolved", or "MarxanSolved" objects.}
}
\value{
"logical" are the objects based on the same data?
}
\description{
This function checks objects to see if they share the same input data.
}
\seealso{
\code{link{MarxanData-class}}, \code{link{MarxanUnsolved-class}}, \code{link{MarxanSolved-class}}
}

