% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{is.gdalInstalled}
\alias{is.gdalInstalled}
\title{Test if GDAL is installed on computer}
\usage{
is.gdalInstalled()
}
\value{
Logical. Is it installed?
}
\description{
This function tests if gdal is installed on the computer.
If not, download it here: \code{\url{http://download.osgeo.org/gdal}}
}
\examples{
gdal_setInstallation()
}
\seealso{
\code{\link[gdalUtils]{gdal_setInstallation}}
}

