% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MarxanResults.R, R/MarxanSolved.R
\name{log}
\alias{log}
\alias{log.MarxanResults}
\alias{log.MarxanSolved}
\title{Log}
\usage{
log(x, ...)

\method{log}{MarxanResults}(x)

\method{log}{MarxanSolved}(x)
}
\arguments{
\item{x}{"MarxanResults" or "MarxanSolved" object.}

\item{...}{not used.}
}
\description{
This function returns the log file associated with running Marxan.
}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved}}, \code{\link{marxan}}
}

