% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{log.MarxanResults}
\alias{log}
\alias{log.MarxanResults}
\alias{log.MarxanSolved}
\title{Log}
\usage{
\method{log}{MarxanResults}(x)

\method{log}{MarxanSolved}(x)

log(x, ...)
}
\arguments{
\item{x}{"MarxanResults" or "MarxanSolved" object.}

\item{...}{not used.}
}
\description{
This function returns the log file associated with running Marxan.
}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved}}, \code{\link{marxan}}
}

