% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/marxan.R
\name{marxan}
\alias{marxan}
\title{General Marxan Function}
\usage{
marxan(..., solve = TRUE)
}
\arguments{
\item{...}{arguments are passed to MarxanData and MarxanOpts functions}

\item{solve}{"logical" should the problem be solved using Marxan?}
}
\value{
"MarxanSolved"  or "MarxanUnsolved"
}
\description{
This is a general function to create Marxan objects from scratch and run the Marxan program to generate solutions.
}
\seealso{
\code{\link{MarxanOpts}}, \code{\link{MarxanData}}, \code{\link{MarxanResults}}, \code{\link{MarxanUnsolved}}
}

