% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R
\docType{class}
\name{MarxanData-class}
\alias{MarxanData-class}
\title{MarxanData: An S4 class to represent Marxan input data}
\description{
This class is used to store Marxan input data.
}
\section{Slots}{

\describe{
\item{\code{polygons}}{"PolySet" planning unit spatial data}

\item{\code{pu}}{"data.frame" planning unit data; with "id", "cost", "status" columns}

\item{\code{species}}{"data.frame" with species data; with "id", "target", "spf", and "name" columns}

\item{\code{puvspecies}}{"data.frame" with data on species density in each planning unit; with "species", "pu", and "target" columns}

\item{\code{boundary}}{"data.frame" with data on the shared boundary length of planning; with "id1", "id2", and "amount" columns}

\item{\code{.cache}}{"environment" used to store processing calculations}
}}

