% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R
\name{MarxanData}
\alias{MarxanData}
\title{Create new MarxanData object}
\usage{
MarxanData(pu, species, puvspecies, boundary, polygons = NULL)
}
\arguments{
\item{pu}{"data.frame" planning unit data; with columns "id", "cost", and "status" columns}

\item{species}{species "data.frame" species data; with columns "id", "target", "spf", and "name" columns}

\item{puvspecies}{puvspecies "data.frame" pu vs. species data; with "species", "pu", and "amount" columns}

\item{boundary}{boundary "data.frame" with shared boundary data; with "id1", "id2", and "boundary" columns}

\item{polygons}{"PolyData" with planning unit data or "NULL" if geoplotting capabilities are not required}
}
\value{
MarxanData object
}
\description{
This function creates a MarxanData using pre-processed data.
}
\examples{
data(planningunits, species, speciesRanges)
x<-MarxanData(
	planningunits@data,
	speciesTargets,
 calcPuVsSpeciesData(planningunits, species),
 calcBoundaryData(planningunits),
	Polygons2PolyData(planning units)
)
}
\seealso{
\code{\link{read.MarxanData}}, \code{\link{write.MarxanData}}, \code{\link{format.MarxanData}}, \code{\link{MarxanData-class}}
}

