% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarxanData.R
\name{MarxanData}
\alias{MarxanData}
\title{Create new MarxanData object}
\usage{
MarxanData(pu, species, puvspecies, boundary, polygons = NULL,
  puvspecies_spo = NULL, skipchecks = FALSE, .cache = new.env(), ...)
}
\arguments{
\item{pu}{"data.frame" planning unit data; with "integer" 'id', "numeric" 'cost', "integer" 'status' columns.}

\item{species}{"data.frame" with species data; with "integer" 'id', "numeric" 'target', "numeric" 'spf', and "character" 'name' columns.}

\item{puvspecies}{"data.frame" with data on species density in each planning unit, with "integer" 'species', "integer" 'pu', and "numeric" 'target' columns. This "data.frame" is sorted in order of 'pu' column.}

\item{boundary}{"data.frame" with data on the shared boundary length of planning; with "integer" 'id1', "integer" 'id2', and "numeric" 'amount' columns.}

\item{polygons}{"PolySet" planning unit spatial data. Defaults to \code{NULL} so that Marxan large problems can still be handled.}

\item{puvspecies_spo}{"data.frame" with data in \code{puvspecies} ordered by 'species' column. Defaults to \code{NULL}, and will generate data from \code{puvspecies}.}

\item{skipchecks}{"logical" Skip data integrity checks? May improve speed for big data sets. Defaults to \code{FALSE}.}

\item{.cache}{"environment" used to cache calculations. Defaults to a new environment.}

\item{...}{not used.}
}
\value{
MarxanData object
}
\description{
This function creates a "MarxanData" object using pre-processed data.
}
\note{
Generally, users are not encouraged to change arguments to \code{.cache}.
}
\examples{
data(taspu, tasinvis)
x<-MarxanData(
	taspu@data,
	data.frame(id=1L, target=12, spf=23, name='species1', stringsAsFactors=FALSE),
 calcPuVsSpeciesData(taspu, tasinvis[[1]]),
 calcBoundaryData(taspu),
 SpatialPolygons2PolySet(taspu)
)
}
\seealso{
\code{\link[PBSmapping]{PolySet}}, \code{\link{read.MarxanData}}, \code{\link{write.MarxanData}}, \code{\link{format.MarxanData}}, \code{\link{MarxanData-class}}.
}

