% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanOpts.R
\name{MarxanOpts}
\alias{MarxanOpts}
\title{Create MarxanOpts object}
\usage{
MarxanOpts(...)
}
\arguments{
\item{BLM}{"numeric" boundary length modifier (default=1000000)}

\item{PROP}{"numeric" proportion of planning units in initial reserve system (default=-1)}

\item{NUMREPS}{"integer" number of replicate runs (default=100)
Annealing parameters}

\item{NUMITNS}{"integer" mumber of iterations for annealing (default=1000000)}

\item{STARTTEMP}{"numeric" initial temperature for annealing (default=-1)}

\item{COOLFAC}{"numeric" cooling factor for annealing (default=0)}

\item{NUMTEMP}{"integer" number of temperature decreases for annealing (default=100000)
Cost threshold}

\item{COSTTHRESH}{"numeric" cost threshold (default=0)}

\item{THRESHPEN1}{"numeric" size of cost threshold penalty (default=0.14)}

\item{THRESHPEN2}{"numeric" shape of cost threshold penalty (default=0)
Program control}

\item{MISSLEVEL}{"numeric" amount of target below which it is counted as 'missing' (default=1)}

\item{ITIMPTYPE}{"integer" iterative improvement type (default=1)}

\item{HEURTYPE}{"integer" heuristic type (default=0)}

\item{CLUMPTYPE}{"integer" clumping penalty type (default=0)}

\item{VERBOSITY}{"integer" amount of output displayed on the program screen (default=3)}

\item{NCORES}{"integer "number of cores to use for processing (default=1)}
}
\value{
"MarxanOpts" object
}
\description{
This function creates a new MarxanOpts object.
}
\details{
General parameters
}
\examples{
x<-MarxanOpts(NCORES=4, NUMREPS=2, NUMITNS=5)
}
\seealso{
\code{\link{MarxanOpts-class}},  \code{\link{read.MarxanOpts}}, \code{\link{write.MarxanOpts}}
}

