% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R
\name{MarxanResults}
\alias{MarxanResults}
\title{Create MarxanResults object}
\usage{
MarxanResults(summary, selections, amountheld, occheld, sepacheived, targetsmet,
  mpm, log)
}
\arguments{
\item{summary}{"data.frame" with summary information on solutions (output_sum.dat).}

\item{selections}{"matrix" with binary selections (output_solutionsmatrix.csv).}

\item{amountheld}{"matrix" with the amount held for each species in each solution (collation of output_mv*.csv files).}

\item{occheld}{"matrix" with the number of occurrences for each species in each solution (collation of output_mv*.csv files).}

\item{sepacheived}{"matrix" indicating the separation achieved (collation of output_mv*.csv files).}

\item{targetsmet}{"matrix" indicating whether the targets have been met for each species in each solution (collation of output_mv*.csv files).}

\item{mpm}{"matrix" indicating minimum proportion met (MPM; collation of output_mv*.csv files).}

\item{log}{"character" with Marxan log  file (otuput_log.dat).}
}
\value{
MarxanResults object
}
\description{
This function creates a new MarxanResults object.
}
\note{
"integer" slot 'best' is automatically calculated from \code{summary}.
}
\seealso{
\code{\link{MarxanResults-class}} \code{\link{read.MarxanResults}}
}

