% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R
\name{MarxanResults}
\alias{MarxanResults}
\title{Create MarxanResults object}
\usage{
MarxanResults(summary, selections, amounthend, occheld, targetsmet, log)
}
\arguments{
\item{summary}{"data.frame" with summary information on solutions}

\item{selections}{"matrix" with binary selections}

\item{occheld}{"matrix" with the number of occurrences for each species in each solution}

\item{log}{"character" with Marxan log file}

\item{dir}{directory from which to load Marxan data}

\item{amountheld}{"matrix" with the amount held for each species in each solution}

\item{best}{"integer" with index of best solution}
}
\value{
MarxanResults object
}
\description{
This function generates a MarxanResults object. Methods are provided to do this using a directory containing Marxan outputs or using objects already loaded into R.
}
\seealso{
\code{\link{MarxanResults-class}} \code{\link{read.MarxanResults}}
}

