% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanSolved.R
\name{MarxanSolved}
\alias{MarxanSolved}
\title{Create new MarxanSolved object}
\usage{
MarxanSolved(MarxanUnsolved, MarxanResults)
}
\arguments{
\item{unsolved}{"MarxanUnsolved" object.}

\item{results}{"MarxanResults" object.}
}
\value{
"MarxanUnsolved" object.
}
\description{
This function creates a MarxanSolved object using a "MarxanUnsolved" object and a "MarxanResults" object.
}
\seealso{
\code{\link{MarxanUnsolved-class}}, \code{\link{MarxanResults-class}}
}

