% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R, R/MarxanSolved.R, R/MarxanUnsolved.R, R/generics.R
\name{maxtargets.MarxanData}
\alias{maxtargets}
\alias{maxtargets.MarxanData}
\alias{maxtargets.MarxanSolved}
\alias{maxtargets.MarxanUnsolved}
\alias{maxtargets<-.MarxanData}
\alias{maxtargets<-.MarxanSolved}
\alias{maxtargets<-.MarxanUnsolved}
\title{Maximum Targets}
\usage{
\method{maxtargets}{MarxanData}(x)

\method{maxtargets}{MarxanData}(x) <- value

\method{maxtargets}{MarxanSolved}(x)

\method{maxtargets}{MarxanSolved}(x) <- value

\method{maxtargets}{MarxanUnsolved}(x)

\method{maxtargets}{MarxanUnsolved}(x) <- value

maxtargets(x, ...)
}
\arguments{
\item{x}{any "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.}
}
\description{
This function returns or assigns the maximum targets for species in a Marxan object. Maximum targets are necessary for setting percent based targets.
}
\section{Methods (by class)}{
\itemize{
\item \code{MarxanData}: 

\item \code{MarxanData}: 

\item \code{MarxanSolved}: 

\item \code{MarxanSolved}: 

\item \code{MarxanUnsolved}: 

\item \code{MarxanUnsolved}: 
}}
\note{
This generic method does not work on "MaranResults" objects because they do not store this information.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

