% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MarxanResults.R, R/MarxanSolved.R
\name{mds}
\alias{mds}
\alias{mds.MarxanResults}
\alias{mds.MarxanSolved}
\title{Metric Dimensional Scaling for Marxan Solutions}
\usage{
mds(x, ...)

\method{mds}{MarxanResults}(x, var = "selections", method = "bray", ...,
  force_reset = FALSE)

\method{mds}{MarxanSolved}(x, var = "selections", method = "bray", ...,
  force_reset = FALSE)
}
\arguments{
\item{x}{"MarxanResults" or "MarxanSolved" object.}

\item{...}{additional arguments to \code{\link[vegan]{monoMDS}}.}

\item{var}{"character" should solutions be compared based on selections ('solutions'), or the amount held ('amountheld'), number of occurances ('occheld'), or whether the targets have been met for each species ('targetsmet')?}

\item{method}{"character" name of distance metric to use for calculating distance (see \code{\link[vegan]{vegdist}}).}

\item{force_reset}{"logical" should analysis be run even if it is stored in the cache?}
}
\value{
"monoMDS" object.
}
\description{
This function runs an metric dimensional scaling (using \code{\link[vegan]{monoMDS}}) on Marxan solutions using various characteristics.
Results cached to permit rapid display for plotting functions.
}
\seealso{
\code{\link{MarxanSolved-class}}, \code{\link{marxan}}, \code{\link{ordiplot}}, \code{\link{dendrogram}}
}

