% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{mpm.MarxanResults}
\alias{mpm}
\alias{mpm.MarxanResults}
\alias{mpm.MarxanSolved}
\title{Extract minimum proportion met ratio}
\usage{
\method{mpm}{MarxanResults}(x, y = NULL)

\method{mpm}{MarxanSolved}(x, y = NULL)

mpm(x, ...)
}
\arguments{
\item{x}{"MarxanResults or "MarxanSolved" object.}

\item{y}{"missing" to return all ratios, "integer" 0 to return ratios for best solution, "integer" value greater than 0 for y'th solution ratios.}
}
\value{
"matrix" or "numeric" vector with solution ratios depending on arguments.
}
\description{
Extract minimum proportion met (MPM) ratio from "MarxanResults" or "MarxanSolved" object.
}
\section{Methods (by class)}{
\itemize{
\item \code{MarxanResults}: 

\item \code{MarxanSolved}: 
}}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved-class}}, \code{\link{marxan}}
}

