% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R, R/MarxanSolved.R, R/MarxanUnsolved.R, R/generics.R
\name{names.MarxanData}
\alias{names}
\alias{names.MarxanData}
\alias{names.MarxanSolved}
\alias{names.MarxanUnsolved}
\alias{names<-.MarxanData}
\alias{names<-.MarxanSolved}
\alias{names<-.MarxanUnsolved}
\title{Species names}
\usage{
\method{names}{MarxanData}(x)

\method{names}{MarxanData}(x) <- value

\method{names}{MarxanSolved}(x)

\method{names}{MarxanSolved}(x) <- value

\method{names}{MarxanUnsolved}(x)

\method{names}{MarxanUnsolved}(x) <- value
}
\arguments{
\item{x}{any "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.}
}
\description{
This function returns or assigns the species names for a Marxan object.
}
\note{
This generic method does not work on "MaranOpts" or "MaranResults" objects because they do not store this information.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

