% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{occheld.MarxanResults}
\alias{occheld}
\alias{occheld.MarxanResults}
\alias{occheld.MarxanSolved}
\title{Extract occurrence held for a solution}
\usage{
\method{occheld}{MarxanResults}(x, y = NULL)

\method{occheld}{MarxanSolved}(x, y = NULL)

occheld(x, ...)

occheld(x, ...)
}
\arguments{
\item{x}{"MarxanResults" or "MarxanSolved" object.}

\item{x}{"MarxanResults" or "MarxanSolved" object.}
}
\value{
"matrix" or "numeric" vector depending on arguments.

"matrix" or "logical" vector depending on arguments.
}
\description{
This function returns the number of occurrences held for each species in a solution.

This function reports whether a solution has met the targets for each species in a solution.
}
\section{Methods (by class)}{
\itemize{
\item \code{MarxanResults}: 

\item \code{MarxanSolved}: 
}}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved}}, \code{\link{marxan}}

\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved}}, \code{\link{marxan}}
}

