% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{occheld.MarxanResults}
\alias{occheld}
\alias{occheld.MarxanResults}
\alias{occheld.MarxanSolved}
\title{Extract occurrence held for a solution}
\usage{
\method{occheld}{MarxanResults}(x, y = NULL)

\method{occheld}{MarxanSolved}(x, y = NULL)

occheld(x, ...)
}
\arguments{
\item{x}{"MarxanResults" or "MarxanSolved" object.}

\item{y}{"NULL" to return all values, "integer" 0 to return values for best solution, "integer" value greater than 0 for \code{y}'th solution value.}
}
\value{
"matrix" or "numeric" vector depending on arguments.
}
\description{
This function returns the number of occurrences held for each species in a solution.
}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved}}, \code{\link{marxan}}
}

