% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{pca.MarxanResults}
\alias{pca}
\alias{pca.MarxanResults}
\alias{pca.MarxanSolved}
\title{Principle components Analysis on Marxan solutions}
\usage{
\method{pca}{MarxanResults}(x, var, ..., force_reset = FALSE)

\method{pca}{MarxanSolved}(x, var = "selections", ..., force_reset = FALSE)

pca(x, ...)
}
\arguments{
\item{x}{"MarxanResults" or "MarxanSolved" object.}

\item{var}{"character" should solutions be compared based on selections ('solutions'), or the amount held ('amountheld'), number of occurances ('occheld'), or whether the targets have been met for each species ('targetsmet')?}

\item{...}{arguments to \code{\link[stats]{prcomp}}.}

\item{force_reset}{"logical" should analysis be run even if it is stored in the cache?}
}
\value{
"prcomp" object with Euclidean distances between rotated data for ordination plots.
}
\description{
This function runs an principle components analysis on Marxan solutions using various characteristics.
Results cached to permit rapid display for plotting functions.
}
\seealso{
\code{\link{MarxanSolved-class}}, \code{\link{marxan}}, \code{\link{ordiplot}}, \code{\link{dendrogram}}
}

