% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{plot}
\alias{plot}
\title{Plot Marxan Solutions}
\usage{
plot(x, y, ...)
}
\arguments{
\item{x}{"MarxanData", "MarxanUnsolved", "MarxanSolved" object.}

\item{y}{See below for details:
    \itemize{
    \item if "missing": function plots the selection frequency of planning units for all solutions.
    \item if "integer": function plots the selection for the n'th solution, use 0 to plot the best solution.
    \item if "character": set to name of species to display its density, or "all" for sum density of all species, or "rich" to get species richness in each planning unit.
    \item if "MarxanSolved" plots the difference in solutions:
    \item \itemize{
        \item if i is "NULL": differences in selection frequencies are plotted.
        \item if i is "integer" and j is "integer", plots differences selection status for solution i in x, and solution j in y. Set i or j to 0 to refer to the best solution in x or y respectively.
        }
}}

\item{colramp}{"character" name of colour palette (see \code{\link[RColorBrewer]{brewer.pal.info}}).}

\item{alpha}{"numeric" alpha value}

\item{basemap}{"character" name of basemap to display. Valid names are "roadmap", "mobile", "satellite", "terrain", "hybrid", "mapmaker-roadmap", "mapmaker-hybrid".}

\item{grayscale}{"logical" should basemap be gray scale?}

\item{xzoom}{"numeric" zoom the geoplot in or out along the x-axis by this percent.}

\item{yzoom}{"numeric" zoom the geoplot in or out along the y-axis by this percent.}

\item{force_reset}{"logical" ignore data in cache? Setting this as ignore will make function slower but may avoid bugs in cache system.}
}
\description{
This function makes a geoplot displaying Marxan solutions.
}
\note{
This function will return an error if spatial polygons were not supplied during the construction of the Marxan object. Furthermore, "MarxanData" and "MarxanUnsolved" objects can only be used to display species densities.
}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved-class}}, \code{\link{basemap}} \code{\link{marxan}}, \code{\link{dendrogram}}, \code{\link{dotchart}}, \code{\link{ordiplot}}
}

