% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarxanData.R
\name{pu}
\alias{pu}
\title{Update Marxan planning unit parameters}
\usage{
pu(id, cost = NA, status = NA)
}
\arguments{
\item{id}{"integer" id of the planning unit.}

\item{cost}{"numeric" new cost value.}

\item{status}{"numeric" new status value.}
}
\value{
"MarxanPuOperation" object.
}
\description{
This function is used in the formula argument of the update function to change planning unit parameters of a "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.
}
\note{
Only specified parameters are changed, all unspecific parameters will remain unchanged.
}
\examples{
pu(1, cost=3)
pu(1, status=1)
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}} \code{\link{update}}, \code{\link{opt}}, \code{\link{spp}}
}

