% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MarxanData.R, R/MarxanUnsolved.R, R/MarxanSolved.R
\name{puids}
\alias{puids}
\alias{puids.MarxanData}
\alias{puids.MarxanSolved}
\alias{puids.MarxanUnsolved}
\alias{puids<-}
\alias{puids<-.MarxanData}
\alias{puids<-.MarxanSolved}
\alias{puids<-.MarxanUnsolved}
\title{Planning unit identifiers}
\usage{
puids(x, ...)

puids(x, ...) <- value

\method{puids}{MarxanData}(x)

\method{puids}{MarxanData}(x) <- value

\method{puids}{MarxanUnsolved}(x)

\method{puids}{MarxanUnsolved}(x) <- value

\method{puids}{MarxanSolved}(x)

\method{puids}{MarxanSolved}(x) <- value
}
\arguments{
\item{x}{any "MarxanOpts", "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.}

\item{...}{not used.}

\item{value}{"integer" new planning unit ids.}
}
\description{
This function returns or assigns the planning unit ids for a Marxan object.
}
\note{
This generic method does not work on "MaranResults" objects because they do not store this information.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

