% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{rasterize.gdal}
\alias{rasterize.gdal}
\title{Rasterize polygon data using GDAL}
\usage{
rasterize.gdal(x, y, ...)
}
\arguments{
\item{x}{"SpatialPolygonsDataFrame" with polygon spatial data}

\item{y}{"RasterLayer" with dimensions, extent, and resolution to be used as a template}

\item{field}{"character" column name with values to burn into the output raster}
}
\value{
RasterLayer
}
\description{
This function converts a SpatialPolygonsDataFrame to raster format using GDAL.
It is expected to be faster than \code{\link[raster]{rasterize}} for large datasets.
However, it will be significantly slower for small datasets.
}
\seealso{
\code{\link[raster]{rasterize}}, \code{\link{is.gdalInstalled}}
}

