% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R
\name{read.MarxanData}
\alias{read.MarxanData}
\title{Read Marxan input}
\usage{
read.MarxanData(path, skipchecks = FALSE)
}
\arguments{
\item{path}{"character" file path for input parameter file or directory containing Marxan files named 'spec.dat', 'pu.dat', 'bound.dat', 'puvspr.dat', and (optionally) 'puvspr_sporder.dat'.}
}
\value{
"MarxanData" object
}
\description{
This function loads Marxan input data from disk.
}
\seealso{
\code{\link{write.MarxanData}}, \code{\link{format.MarxanData}}, \code{\link{MarxanData}}, \code{\link{MarxanData-class}}.
}

