% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{score.MarxanResults}
\alias{score}
\alias{score.MarxanResults}
\alias{score.MarxanSolved}
\title{Extract Solution Score}
\usage{
\method{score}{MarxanResults}(x, y = "best")

\method{score}{MarxanSolved}(x, y = "best")

score(x, ...)
}
\arguments{
\item{x}{"MarxanResults or "MarxanSolved" object.}

\item{y}{'best' to return best solution score, 'all' to return all scores, or "integer" to return score for y'th solution.}
}
\value{
"matrix" or "numeric" vector with solution scores depending on arguments.
}
\description{
Extract solution score from "MarxanResults" or "MarxanSolved" object.
}
\section{Methods (by class)}{
\itemize{
\item \code{MarxanResults}: 

\item \code{MarxanSolved}: 
}}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved-class}}, \code{\link{marxan}}
}

