% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MarxanResults.R, R/MarxanSolved.R
\name{score}
\alias{score}
\alias{score.MarxanResults}
\alias{score.MarxanSolved}
\title{Extract solution score}
\usage{
score(x, ...)

\method{score}{MarxanResults}(x, y = NULL)

\method{score}{MarxanSolved}(x, y = NULL)
}
\arguments{
\item{x}{"MarxanResults or "MarxanSolved" object.}

\item{...}{not used.}

\item{y}{"NULL" to return all scores, "integer" 0 to return score for best solution, "integer" value greater than 0 for \code{y}'th solution score.}
}
\value{
"matrix" or "numeric" vector with solution score(s) depending on arguments.
}
\description{
Extract solution score from "MarxanResults" or "MarxanSolved" object.
}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved-class}}, \code{\link{marxan}}
}

