% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{selection.MarxanResults}
\alias{selection}
\alias{selection.MarxanResults}
\alias{selection.MarxanSolved}
\title{Extract solution selection}
\usage{
\method{selection}{MarxanResults}(x, y = NULL)

\method{selection}{MarxanSolved}(x, y = NULL)

selection(x, ...)
}
\arguments{
\item{x}{"MarxanResults or "MarxanSolved" object.}

\item{y}{"missing" to return all solution selections, "integer" 0 to return selection for best solution, "integer" value greater than 0 for y'th solution selection.}
}
\value{
"matrix" or "numeric" vector with planning units statuses depending on arguments.
}
\description{
Extract selections for a given solution from a "MarxanResults or "MarxanSolved" object.
}
\section{Methods (by class)}{
\itemize{
\item \code{MarxanResults}: 

\item \code{MarxanSolved}: 
}}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved-class}}, \code{\link{marxan}}.
}

