% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MarxanResults.R, R/MarxanSolved.R
\name{sepacheived}
\alias{sepacheived}
\alias{sepacheived.MarxanResults}
\alias{sepacheived.MarxanSolved}
\title{Extract separation achieved value}
\usage{
sepacheived(x, ...)

\method{sepacheived}{MarxanResults}(x, y = NULL)

\method{sepacheived}{MarxanSolved}(x, y = NULL)
}
\arguments{
\item{x}{"MarxanResults or "MarxanSolved" object.}

\item{...}{not used.}

\item{y}{"NULL" to return all values, "integer" 0 to return values for best solution, "integer" value greater than 0 for \code{y}'th solution values.}
}
\value{
"matrix" or "numeric" vector with values depending on arguments.
}
\description{
Extract separation achieved value from "MarxanResults" or "MarxanSolved" object.
}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved-class}}, \code{\link{marxan}}
}

