% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Polygons2PolySet.R
\name{SpatialPolygons2PolySet}
\alias{SpatialPolygons2PolySet}
\title{Convert "SpatialPolygons" to "PolySet" data}
\usage{
SpatialPolygons2PolySet(x, n=10000L)
}
\arguments{
\item{x}{"SpatialPolygons" or "SpatialPolygonsDataFrame" object.}

\item{n_preallocate}{"integer" How much memory should be preallocated for processing? Ideally, this number should equal the number of vertices in the "SpatialPolygons" object. If dtaa processing is taking too long consider increasing this value.}
}
\value{
"PolySet" object.
}
\description{
This function converts spatial objects "SpatialPolygons" and "SpatialPolygonsDataFrame" to "PolySet".
}
\note{
Be aware that this function is designed to be as fast as possible, but as a result it depends on C++ code and if used inappropriately this function will crash R.
}
\examples{
data(planningunits)
x <- SpatialPolygons2PolySet(planningunits)
}
\seealso{
For a slower, more stable equivalent see \code{\link[maptools]{SpatialPolygons2PolySet}}.
}

