% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R, R/MarxanSolved.R, R/MarxanUnsolved.R, R/generics.R
\name{spfs.MarxanData}
\alias{spfs}
\alias{spfs.MarxanData}
\alias{spfs.MarxanSolved}
\alias{spfs.MarxanUnsolved}
\alias{spfs<-}
\alias{spfs<-.MarxanData}
\alias{spfs<-.MarxanSolved}
\alias{spfs<-.MarxanUnsolved}
\title{Species penalty factors}
\usage{
\method{spfs}{MarxanData}(x)

\method{spfs}{MarxanData}(x) <- value

\method{spfs}{MarxanSolved}(x)

\method{spfs}{MarxanSolved}(x) <- value

\method{spfs}{MarxanUnsolved}(x)

\method{spfs}{MarxanUnsolved}(x) <- value

spfs(x, ...)

spfs(x, ...) <- value
}
\arguments{
\item{x}{any "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.}
}
\description{
This function returns or assigns the species penalty factors for speces in a Marxan object.
}
\note{
This generic method does not work on "MaranResults" objects because they do not store this information.
}
\seealso{
\code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

