% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MarxanData.R, R/MarxanUnsolved.R, R/MarxanSolved.R
\name{sppids}
\alias{sppids}
\alias{sppids.MarxanData}
\alias{sppids.MarxanSolved}
\alias{sppids.MarxanUnsolved}
\alias{sppids<-}
\alias{sppids<-.MarxanData}
\alias{sppids<-.MarxanSolved}
\alias{sppids<-.MarxanUnsolved}
\title{Species identifiers}
\usage{
sppids(x, ...)

sppids(x, ...) <- value

\method{sppids}{MarxanData}(x)

\method{sppids}{MarxanData}(x) <- value

\method{sppids}{MarxanUnsolved}(x)

\method{sppids}{MarxanUnsolved}(x) <- value

\method{sppids}{MarxanSolved}(x)

\method{sppids}{MarxanSolved}(x) <- value
}
\arguments{
\item{x}{any "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.}

\item{...}{not used.}

\item{value}{"integer" new species ids.}
}
\description{
This function returns or assigns the species ids for a Marxan object.
}
\note{
This generic method does not work on "MaranResults" objects because they do not store this information.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

