% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R, R/MarxanSolved.R, R/MarxanUnsolved.R, R/generics.R
\docType{methods}
\name{spplot,MarxanData-method}
\alias{spplot}
\alias{spplot,MarxanData-method}
\alias{spplot,MarxanSolved-method}
\alias{spplot,MarxanUnsolved-method}
\title{Plot species distribution in planning units}
\usage{
\S4method{spplot}{MarxanData}(obj, y = obj@data@species$id, var = "amount",
  basemap = "none", colramp = "YlOrRd", alpha = ifelse(basemap == "none",
  1, 0.7), grayscale = FALSE, force_reset = FALSE)

\S4method{spplot}{MarxanSolved}(obj, y = obj@data@species$id,
  var = "amount", basemap = "none", colramp = "YlOrRd",
  alpha = ifelse(basemap == "none", 1, 0.4), grayscale = FALSE,
  force_reset = FALSE)

\S4method{spplot}{MarxanUnsolved}(obj, y = obj@data@species$id,
  var = "amount", basemap = "none", colramp = "YlOrRd",
  alpha = ifelse(basemap == "none", 1, 0.7), grayscale = FALSE,
  force_reset = FALSE)

spplot(obj, ...)
}
\arguments{
\item{obj}{"MarxanData", "MarxanUnsolved", "MarxanSolved" object.}

\item{y}{"character" name of species, or "integer" species id.}

\item{var}{"character" \itemize{
\item if 'amount': sum density of \code{y} species in planning units is shown.
\item if 'occ': number of \code{y} species in planning units is shown.
}}

\item{basemap}{"character" name of google base map. Defaults to "none".}

\item{colramp}{"character" name of color ramp. See color palettes available in \code{\link[RColorBrewer]{brewer.pal.ino}}}

\item{alpha}{"numeric" between 0 and 1 describing how transparent the colors should be. Defaults to 1 if \code{basemap}='none', otherwise defaults to 0.7.}

\item{grayscale}{"logical" Shoould be basemap be gray scaled?}

\item{force_reset}{"logical" should base map image be downloaded even if it is stored in the cache?}
}
\description{
This function plots the density of species in planning units in "Marxan" objects.
}
\seealso{
\code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanData-class}}, \code{\link[RgoogleMaps]{GetMap}}, \code{\link[RgoogleMaps]{PlotOnStaticMap}}
}

