% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanData.R, R/MarxanSolved.R, R/MarxanUnsolved.R, R/generics.R
\name{targets.MarxanData}
\alias{targets}
\alias{targets.MarxanData}
\alias{targets.MarxanSolved}
\alias{targets.MarxanUnsolved}
\alias{targets<-.MarxanData}
\alias{targets<-.MarxanSolved}
\alias{targets<-.MarxanUnsolved}
\title{Targets}
\usage{
\method{targets}{MarxanData}(x)

\method{targets}{MarxanData}(x) <- value

\method{targets}{MarxanSolved}(x)

\method{targets}{MarxanSolved}(x) <- value

\method{targets}{MarxanUnsolved}(x)

\method{targets}{MarxanUnsolved}(x) <- value

targets(x, ...)
}
\arguments{
\item{x}{any "MarxanOpts", "MarxanData", "MarxanUnsolved", or "MarxanSolved" object.}
}
\description{
This function returns or assigns the species penalty factors for a Marxan object.
}
\section{Methods (by class)}{
\itemize{
\item \code{MarxanData}: 

\item \code{MarxanData}: 

\item \code{MarxanSolved}: 

\item \code{MarxanSolved}: 

\item \code{MarxanUnsolved}: 

\item \code{MarxanUnsolved}: 
}}
\note{
This generic method does not work on "MaranResults" objects because they do not store this information.
}
\seealso{
\code{\link{MarxanOpts-class}}, \code{\link{MarxanData-class}}, \code{\link{MarxanUnsolved-class}}, \code{\link{MarxanSolved-class}}
}

