% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanResults.R, R/MarxanSolved.R, R/generics.R
\name{targetsmet.MarxanResults}
\alias{targetsmet}
\alias{targetsmet.MarxanResults}
\alias{targetsmet.MarxanSolved}
\title{Extract information on whether solutions have met targets}
\usage{
\method{targetsmet}{MarxanResults}(x, y = NULL)

\method{targetsmet}{MarxanSolved}(x, y = NULL)

targetsmet(x, ...)
}
\arguments{
\item{x}{"MarxanResults" or "MarxanSolved" object.}

\item{y}{"NULL" to return all values, "integer" 0 to return values for best solution, "integer" value greater than 0 for \code{y}'th solution value.}
}
\value{
"matrix" or "logical" vector depending on arguments.
}
\description{
This function reports whether a solution has met the targets for each species in a solution.
}
\seealso{
\code{\link{MarxanResults-class}}, \code{\link{MarxanSolved}}, \code{\link{marxan}}
}

