% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tasinvis}
\alias{tasinvis}
\title{Broad vegetation data}
\format{An object of class \code{"RasterLayer"}.}
\usage{
data(tasinvis)
}
\description{
"RasterStack" data for the distribution of 63 vegetation classes in Tasmania, Australia. Each 'RasterLayer' in the 'RasterStack' represents the binary distribution for a single vegetation class. This data was obtained from the "Introduction to Marxan" course.
}
\examples{
data(tasinvis)
plot(tasinvis)
}
\seealso{
\code{\link{taspu}}
}
\keyword{datasets}

