% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MarxanOpts.R
\name{write.MarxanOpts}
\alias{write.MarxanOpts}
\title{Write Marxan Input Parameters to Disk}
\usage{
write.MarxanOpts(x, dir, outputdir = dir, seed = sample.int(n = 10000L, n =
  1L))
}
\arguments{
\item{dir}{"character" directory path for location to save input parameters file}

\item{outputdir}{"character" directory path for location where Marxan input data files are saved}

\item{seed}{"integer" seed for random number generation in Marxan}
}
\description{
This function writes Marxan parameter settings to a file.
}
\examples{
x<-MarxanOpts()
write.MarxanData(x, file.path(tempdir(), 'input.dat'))
y<-read.MarxanData(file.path(tempdir(), 'input.dat'))
stopifnot(identical(x,y))
}
\seealso{
\code{\link{MarxanOpts-class}},  \code{\link{MarxanOpts}}, \code{\link{read.MarxanOpts}}
}

