% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zonalSum.R
\name{zonalSum}
\alias{zonalSum}
\title{Zonal Sum}
\usage{
zonalSum(x, y, ...)
}
\arguments{
\item{x}{SpatialPolygons or SpatialPolygons object}

\item{y}{RasterLayer, RasterStack, or RasterBrick}

\item{speciesName}{character vector of species names}

\item{ncores}{Number of cores to use for processing data.}
}
\value{
data.frame with sum of raster values in each polygon.
}
\description{
This function calculates the zonal sum of SpatialPolygons in a RasterLayer, RasterStack, or RasterBrick.
For most cases it is expected to run faster than the zonal functon in the raster package.
This function is useful for extracting the sum of raster values inside each planning unit.
}
\examples{
data(species, planningunits)
zonalSum(planningunits, species[[1]])
zonalSum(planningunits, species)
}
\seealso{
\code{\link[raster]{zonal}}
}

